docs: unify the Acme labs on one engine, fix narration, and test every step
- Chapter 1 now opens with the failure: run the report as Alice, watch
  the error move from the schema gate to the table gate, then open the
  smallest working path.
- Chapter 4 no longer claims deploy needs SET ROLE to pass owner checks
  (an inheriting member of the owner already passes them); the recipe is
  reframed around who owns the objects a migration creates, which is the
  bug chapter 5 then delivers.
- Chapter 5 splits the dense repair step: fix the existing table first,
  then configure defaults and watch them fire on a freshly created table.
- Chapter 7 is ported onto the shared AcmeStoryLab engine with per-step
  seeding and the Acme cast (Bob, team_lead, Dana), replacing the stale
  stateful lab; PostgresPermissionLab.jsx and the unused
  ProductionRoleShapeDiagram are removed.
- The playground database now follows the real retirement sequence, so
  Priya's legacy objects survive under app_owner, and new challenges
  close the Alice and Priya arcs.
- Step gating is advisory instead of blocking: every step seeds its own
  database, so readers can jump anywhere; the canonical run remains the
  reward.
- Lab data moves to a plain module (AcmeStoryData.mjs) shared by the
  React component and a Node test suite that executes every canonical
  step and the narration-critical counterfactuals against PGlite; the
  suite runs in the docs workflow before the build.
