+++ docs/package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build --webpack",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "test:labs": "node tests/labs.test.mjs"
   },
   "browserslist": "defaults, not ie <= 11",
   "dependencies": {
