+++ docs/src/components/AcmeStoryData.mjs
@@ -1,0 +1,1235 @@
+// Canonical data for the Acme story labs: database seeds, per-step SQL,
+// inspection queries, and pass conditions. This module is plain JavaScript on
+// purpose — the React component renders it and `docs/tests/labs.test.mjs`
+// executes every canonical step against the same PGlite build the browser
+// uses, so a PGlite upgrade or a wrong expectation fails CI instead of a
+// reader.
+
+const ordersTable = `CREATE TABLE app.orders (
+  id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
+  customer text NOT NULL,
+  total_cents bigint NOT NULL
+);
+INSERT INTO app.orders (customer, total_cents)
+VALUES ('Acme', 4200), ('Globex', 9900);`;
+
+const founderSeed = `
+CREATE ROLE priya LOGIN;
+CREATE ROLE alice LOGIN;
+CREATE ROLE bob LOGIN;
+CREATE ROLE reporting_app LOGIN;
+CREATE ROLE deploy LOGIN;
+GRANT CREATE ON DATABASE postgres TO priya;
+SET ROLE priya;
+CREATE SCHEMA app AUTHORIZATION priya;
+${ordersTable}
+RESET ROLE;`;
+
+const schemaOnlySeed = `${founderSeed}
+GRANT USAGE ON SCHEMA app TO alice;`;
+
+const directAccessSeed = `${schemaOnlySeed}
+GRANT SELECT ON app.orders TO alice;`;
+
+const capabilitySeed = `${directAccessSeed}
+CREATE ROLE orders_reader;
+GRANT USAGE ON SCHEMA app TO orders_reader;
+GRANT SELECT ON app.orders TO orders_reader;
+GRANT orders_reader TO alice, reporting_app;`;
+
+const postDriftSeed = `${capabilitySeed}
+GRANT orders_reader TO bob;
+REVOKE orders_reader FROM alice;
+REVOKE SELECT ON app.orders FROM alice;
+REVOKE USAGE ON SCHEMA app FROM alice;`;
+
+const ownerSeed = `${postDriftSeed}
+CREATE ROLE app_owner;
+GRANT app_owner TO deploy;
+ALTER SCHEMA app OWNER TO app_owner;
+ALTER TABLE app.orders OWNER TO app_owner;
+ALTER SEQUENCE app.orders_id_seq OWNER TO app_owner;`;
+
+const defaultsSeed = `${ownerSeed}
+SET SESSION AUTHORIZATION deploy;
+CREATE TABLE app.refunds (
+  id bigint PRIMARY KEY,
+  order_id bigint NOT NULL,
+  total_cents bigint NOT NULL
+);
+INSERT INTO app.refunds VALUES (1, 1, 500);
+SET SESSION AUTHORIZATION postgres;`;
+
+const repairedSeed = `${defaultsSeed}
+GRANT SELECT ON app.refunds TO orders_reader;
+ALTER TABLE app.refunds OWNER TO app_owner;`;
+
+const durableSeed = `${repairedSeed}
+ALTER DEFAULT PRIVILEGES FOR ROLE app_owner IN SCHEMA app
+  GRANT SELECT ON TABLES TO orders_reader;
+SET ROLE app_owner;
+CREATE TABLE app.shipments (
+  id bigint PRIMARY KEY,
+  order_id bigint NOT NULL,
+  status text NOT NULL
+);
+INSERT INTO app.shipments VALUES (1, 1, 'sent');
+RESET ROLE;`;
+
+const offboardingSeed = `${durableSeed}
+SET ROLE priya;
+CREATE SCHEMA legacy AUTHORIZATION priya;
+CREATE TABLE legacy.customer_notes (id bigint PRIMARY KEY, note text);
+RESET ROLE;`;
+
+const retiredSeed = `${offboardingSeed}
+REASSIGN OWNED BY priya TO app_owner;
+DROP OWNED BY priya;
+DROP ROLE priya;`;
+
+const mechanicsSeed = `${retiredSeed}
+CREATE ROLE analyst;
+CREATE ROLE team_lead LOGIN;
+CREATE ROLE dana LOGIN;`;
+
+const nestedSeed = `${mechanicsSeed}
+GRANT orders_reader TO analyst;
+GRANT analyst TO bob;
+REVOKE orders_reader FROM bob;`;
+
+const dormantSeed = `${nestedSeed}
+REVOKE analyst FROM bob;
+GRANT analyst TO bob WITH INHERIT FALSE, SET TRUE;`;
+
+const delegatedSeed = `${nestedSeed}
+GRANT analyst TO team_lead WITH ADMIN TRUE, INHERIT FALSE, SET FALSE;`;
+
+const completeSeed = `${nestedSeed}
+GRANT analyst TO team_lead WITH ADMIN TRUE, INHERIT FALSE, SET FALSE;
+GRANT analyst TO dana;
+CREATE ROLE contractor LOGIN;
+CREATE ROLE auditor LOGIN;
+CREATE SCHEMA billing_api AUTHORIZATION app_owner;
+GRANT USAGE ON SCHEMA billing_api TO PUBLIC;
+SET ROLE app_owner;
+CREATE FUNCTION billing_api.order_total()
+RETURNS bigint LANGUAGE sql SECURITY DEFINER
+SET search_path = pg_catalog, app
+AS 'SELECT sum(total_cents)::bigint FROM app.orders';
+RESET ROLE;
+GRANT EXECUTE ON FUNCTION billing_api.order_total() TO auditor;`;
+
+function privilegeInspection(role, object = "app.orders") {
+  return `SELECT
+  has_schema_privilege('${role}', 'app', 'USAGE') AS schema_usage,
+  has_table_privilege('${role}', '${object}', 'SELECT') AS object_select,
+  CASE
+    WHEN EXISTS (SELECT 1 FROM pg_roles WHERE rolname = 'orders_reader')
+    THEN pg_has_role('${role}', 'orders_reader', 'USAGE')
+    ELSE false
+  END AS inherits_reader,
+  (SELECT tableowner FROM pg_tables
+    WHERE schemaname = 'app' AND tablename = '${
+      object.split(".")[1]
+    }') AS object_owner;`;
+}
+
+const basicCards = (role, resultLabel) => (row, output) =>
+  [
+    ["Role", role, "neutral"],
+    [
+      "Membership",
+      row?.inherits_reader ? "orders_reader" : "none",
+      row?.inherits_reader ? "pass" : "neutral",
+    ],
+    [
+      "Schema",
+      row?.schema_usage ? "USAGE" : "blocked",
+      row?.schema_usage ? "pass" : "blocked",
+    ],
+    [
+      "Table",
+      row?.object_select ? "SELECT" : "blocked",
+      row?.object_select ? "pass" : "blocked",
+    ],
+    [
+      "Result",
+      output.error ? "denied" : resultLabel,
+      output.error ? "blocked" : "pass",
+    ],
+  ];
+
+const gateCards = (role, resultLabel) => (row, output) =>
+  [
+    ["Role", role, "neutral"],
+    [
+      "Schema",
+      row?.schema_usage ? "USAGE" : "blocked",
+      row?.schema_usage ? "pass" : "blocked",
+    ],
+    [
+      "Table",
+      row?.object_select ? "SELECT" : "blocked",
+      row?.object_select ? "pass" : "blocked",
+    ],
+    [
+      "Result",
+      output.error ? "denied" : resultLabel,
+      output.error ? "blocked" : "pass",
+    ],
+  ];
+
+const chapters = {
+  gates: {
+    eyebrow: "Chapter 1 · The permission chain",
+    title: "Alice gets the first report working",
+    description:
+      "Acme needs an orders report today. Run the query first, let PostgreSQL name each closed gate, and open exactly the path the report needs.",
+    next: {
+      href: "/docs/postgresql-capability-roles",
+      title: "Replace direct grants with a capability role",
+    },
+    actors: [
+      ["postgres", "Database admin"],
+      ["alice", "Alice"],
+    ],
+    steps: [
+      {
+        title: "Run the report as Alice",
+        why: "The application already reads orders, but nobody has granted Alice anything. PostgreSQL evaluates the query against the current role, gate by gate, and reports the first gate that fails.",
+        prompt: "Run the report query as Alice.",
+        setup: founderSeed,
+        role: "alice",
+        sql: "SELECT * FROM app.orders ORDER BY id;",
+        inspect: privilegeInspection("alice"),
+        cards: gateCards("Alice", "rows"),
+        expect: (output) =>
+          output.error?.includes("permission denied for schema app"),
+        observation:
+          "PostgreSQL stopped at the first closed gate and named it: the schema. It never reached the table check — the error identifies where evaluation stopped, not everything that is missing.",
+      },
+      {
+        title: "Open the schema gate",
+        why: "Schema USAGE makes names inside app reachable. It says nothing about what Alice may do to the objects she can now reach.",
+        prompt: "Grant Alice USAGE on the app schema.",
+        setup: founderSeed,
+        role: "postgres",
+        sql: "GRANT USAGE ON SCHEMA app TO alice;",
+        inspect: privilegeInspection("alice"),
+        cards: gateCards("Alice", "USAGE granted"),
+        expect: (output, row) =>
+          !output.error && row?.schema_usage && !row?.object_select,
+        observation:
+          "The schema gate is open, but the diagram shows the table gate is still closed. Predict the next error before you run the query again.",
+      },
+      {
+        title: "Watch the error move",
+        why: "The same query now gets past the schema. If PostgreSQL evaluates gates in order, the error should move to the next one.",
+        prompt: "Run the identical report query as Alice again.",
+        setup: schemaOnlySeed,
+        role: "alice",
+        sql: "SELECT * FROM app.orders ORDER BY id;",
+        inspect: privilegeInspection("alice"),
+        cards: gateCards("Alice", "rows"),
+        expect: (output) =>
+          output.error?.includes("permission denied for table orders"),
+        observation:
+          "Same query, new error. The failure moved from the schema to the table, which proves the gates are separate: USAGE did not imply SELECT.",
+      },
+      {
+        title: "Authorize the table read",
+        why: "Table SELECT is the last gate between Alice and the rows. Together with schema USAGE it completes the path.",
+        prompt: "Grant Alice SELECT on the orders table.",
+        setup: schemaOnlySeed,
+        role: "postgres",
+        sql: "GRANT SELECT ON app.orders TO alice;",
+        inspect: privilegeInspection("alice"),
+        cards: gateCards("Alice", "path ready"),
+        expect: (output, row) =>
+          !output.error && row?.schema_usage && row?.object_select,
+        observation:
+          "Alice now has a complete path: PostgreSQL can resolve app.orders, then authorize SELECT.",
+      },
+      {
+        title: "Run the real report",
+        why: "Permissions matter only when the application operation succeeds. Run the same SELECT the report will send.",
+        prompt: "Query orders as Alice.",
+        setup: directAccessSeed,
+        role: "alice",
+        sql: "SELECT * FROM app.orders ORDER BY id;",
+        inspect: privilegeInspection("alice"),
+        cards: gateCards("Alice", "2 rows"),
+        expect: (output) =>
+          !output.error && output.results[0]?.rows.length === 2,
+        observation:
+          "The report works. Remember that these convenient direct grants will survive until someone explicitly removes them.",
+      },
+    ],
+  },
+  capabilities: {
+    eyebrow: "Chapter 2 · Capability roles",
+    title: "A reporting application joins Alice",
+    description:
+      "A second consumer needs the same access. Name the capability once instead of copying grants onto every login.",
+    next: {
+      href: "/docs/postgresql-access-drift",
+      title: "Make the old direct grants resurface",
+    },
+    actors: [
+      ["postgres", "Database admin"],
+      ["alice", "Alice"],
+      ["reporting_app", "Reporting app"],
+    ],
+    steps: [
+      {
+        title: "Create orders_reader",
+        why: "The permission bundle is a job, not a person. A NOLOGIN role gives that job a durable name and keeps object ACLs independent from staff changes.",
+        prompt:
+          "Move the reusable grants onto orders_reader and make both consumers members. Deliberately leave Alice’s Chapter 1 grants in place.",
+        setup: directAccessSeed,
+        role: "postgres",
+        sql: `CREATE ROLE orders_reader;
+GRANT USAGE ON SCHEMA app TO orders_reader;
+GRANT SELECT ON app.orders TO orders_reader;
+GRANT orders_reader TO alice, reporting_app;`,
+        inspect: privilegeInspection("reporting_app"),
+        cards: basicCards("Reporting app", "path ready"),
+        expect: (output, row) =>
+          !output.error &&
+          row?.inherits_reader &&
+          row?.schema_usage &&
+          row?.object_select,
+        observation:
+          "One membership now gives the reporting application the same complete path as Alice.",
+      },
+      {
+        title: "Prove the application role works",
+        why: "A policy graph is useful only if the actual login inherits the intended privileges.",
+        prompt: "Run the report as reporting_app.",
+        setup: capabilitySeed,
+        role: "reporting_app",
+        sql: "SELECT * FROM app.orders ORDER BY id;",
+        inspect: privilegeInspection("reporting_app"),
+        cards: basicCards("Reporting app", "2 rows"),
+        expect: (output) =>
+          !output.error && output.results[0]?.rows.length === 2,
+        observation:
+          "The login has no direct table grant. Its membership in orders_reader supplies both gates.",
+      },
+      {
+        title: "Count Alice’s access paths",
+        why: "Adding the cleaner path did not remove the old one. Alice can now read through membership and through the direct grants you created in Chapter 1.",
+        prompt: "Ask PostgreSQL which paths still exist.",
+        setup: capabilitySeed,
+        role: "postgres",
+        sql: `SELECT
+  pg_has_role('alice', 'orders_reader', 'USAGE') AS through_membership,
+  EXISTS (
+    SELECT 1 FROM information_schema.table_privileges
+    WHERE grantee = 'alice' AND table_schema = 'app'
+      AND table_name = 'orders' AND privilege_type = 'SELECT'
+  ) AS direct_table_grant;`,
+        inspect: privilegeInspection("alice"),
+        cards: (row) => [
+          [
+            "Alice path 1",
+            row?.inherits_reader ? "membership" : "missing",
+            row?.inherits_reader ? "pass" : "blocked",
+          ],
+          [
+            "Alice path 2",
+            row?.object_select ? "direct ACL remains" : "missing",
+            row?.object_select ? "focus" : "blocked",
+          ],
+          ["Lesson", "two paths", "focus"],
+        ],
+        expect: (output) =>
+          output.results[0]?.rows[0]?.through_membership === true &&
+          output.results[0]?.rows[0]?.direct_table_grant === true,
+        observation:
+          "This duplicate path is harmless today, but it is already authorization drift waiting for a team change.",
+      },
+    ],
+  },
+  drift: {
+    eyebrow: "Chapter 3 · Drift",
+    title: "Bob joins the team; Alice leaves it",
+    description:
+      "Membership makes the team change look simple—until the direct grants you created earlier defeat the offboarding test.",
+    next: {
+      href: "/docs/postgresql-ownership",
+      title: "Meet PostgreSQL ownership",
+    },
+    actors: [
+      ["postgres", "Database admin"],
+      ["alice", "Alice"],
+      ["bob", "Bob"],
+    ],
+    steps: [
+      {
+        title: "Change the team",
+        why: "The intended rule is now “Bob is an orders reader; Alice is not.” Membership expresses that change in two edges.",
+        prompt: "Add Bob and remove Alice from orders_reader.",
+        setup: capabilitySeed,
+        role: "postgres",
+        sql: `GRANT orders_reader TO bob;
+REVOKE orders_reader FROM alice;`,
+        inspect: privilegeInspection("alice"),
+        cards: (row) => [
+          [
+            "Alice membership",
+            row?.inherits_reader ? "still present" : "removed",
+            row?.inherits_reader ? "blocked" : "pass",
+          ],
+          [
+            "Schema",
+            row?.schema_usage ? "still reachable" : "blocked",
+            row?.schema_usage ? "focus" : "pass",
+          ],
+          [
+            "Table",
+            row?.object_select ? "still readable" : "blocked",
+            row?.object_select ? "focus" : "pass",
+          ],
+        ],
+        expect: (output, row) =>
+          !output.error &&
+          !row?.inherits_reader &&
+          row?.schema_usage &&
+          row?.object_select,
+        observation:
+          "The membership is gone, but Alice’s effective access is not. PostgreSQL can still see the direct Chapter 1 ACLs.",
+      },
+      {
+        title: "Make the failed offboarding visible",
+        why: "Never infer loss of access from one revoked edge. Test the same operation the role used before.",
+        prompt: "Run the report as Alice after her membership was removed.",
+        setup: `${capabilitySeed}\nGRANT orders_reader TO bob;\nREVOKE orders_reader FROM alice;`,
+        role: "alice",
+        sql: "SELECT * FROM app.orders ORDER BY id;",
+        inspect: privilegeInspection("alice"),
+        cards: basicCards("Alice", "still 2 rows"),
+        expect: (output) =>
+          !output.error && output.results[0]?.rows.length === 2,
+        observation:
+          "The query still succeeds. Effective access is the union of every surviving path, not the edge you meant to remove.",
+      },
+      {
+        title: "Remove the undeclared path",
+        why: "The direct schema and table ACLs no longer belong in desired policy. Revoking both makes membership the only path.",
+        prompt: "Remove Alice’s old direct grants.",
+        setup: `${capabilitySeed}\nGRANT orders_reader TO bob;\nREVOKE orders_reader FROM alice;`,
+        role: "postgres",
+        sql: `REVOKE SELECT ON app.orders FROM alice;
+REVOKE USAGE ON SCHEMA app FROM alice;`,
+        inspect: privilegeInspection("alice"),
+        cards: basicCards("Alice", "denied"),
+        expect: (output, row) =>
+          !output.error &&
+          !row?.inherits_reader &&
+          !row?.schema_usage &&
+          !row?.object_select,
+        observation:
+          "Alice now has no surviving path. An authoritative pgroles plan would have exposed these undeclared grants before the team change.",
+      },
+      {
+        title: "Verify the negative case",
+        why: "Offboarding is complete only when the old operation fails and the replacement reader still succeeds.",
+        prompt: "Try the report as Alice. Then switch to Bob and run it again.",
+        setup: postDriftSeed,
+        role: "alice",
+        sql: "SELECT * FROM app.orders ORDER BY id;",
+        inspect: privilegeInspection("alice"),
+        cards: basicCards("Alice", "denied"),
+        expect: (output) =>
+          output.error?.includes("permission denied for schema app"),
+        observation:
+          "The denial is the evidence. Use the role selector to confirm Bob receives the same rows Alice used to receive.",
+      },
+    ],
+  },
+  ownership: {
+    eyebrow: "Chapter 4 · Ownership",
+    title: "The migration runner hits Priya’s old table",
+    description:
+      "The original founder created orders herself. Object grants can allow data access, but only an owner—or a role holding the owner’s privileges—can change the table.",
+    next: {
+      href: "/docs/postgresql-default-privileges",
+      title: "Keep future tables readable",
+    },
+    actors: [
+      ["deploy", "Deploy"],
+      ["postgres", "Database admin"],
+      ["app_owner", "App owner"],
+    ],
+    steps: [
+      {
+        title: "Let deploy try the migration",
+        why: "Deploy can connect, but Acme never gave it ownership of Priya’s table. ALTER TABLE checks ownership rather than an ordinary ACL privilege.",
+        prompt: "Run the migration as deploy.",
+        setup: `${postDriftSeed}
+GRANT USAGE ON SCHEMA app TO deploy;`,
+        role: "deploy",
+        sql: "ALTER TABLE app.orders ADD COLUMN source text;",
+        inspect: `SELECT tableowner AS object_owner FROM pg_tables WHERE schemaname = 'app' AND tablename = 'orders';`,
+        cards: (row, output) => [
+          ["Current role", "Deploy", "neutral"],
+          ["Table owner", row?.object_owner || "unknown", "focus"],
+          [
+            "ALTER TABLE",
+            output.error ? "owner required" : "changed",
+            output.error ? "blocked" : "pass",
+          ],
+        ],
+        expect: (output) =>
+          output.error?.includes("must be owner of table orders"),
+        observation:
+          "No combination of SELECT, INSERT, UPDATE, or DELETE grants makes deploy the owner. Ownership is a separate authority.",
+      },
+      {
+        title: "Replace the human owner",
+        why: "Production objects should outlive people. app_owner is a durable NOLOGIN role. Granting it to deploy hands deploy the owner’s authority through membership—and permission to become the owner deliberately.",
+        prompt:
+          "Create the owner role, let deploy hold it, and transfer the schema and existing objects.",
+        setup: postDriftSeed,
+        role: "postgres",
+        sql: `CREATE ROLE app_owner;
+GRANT app_owner TO deploy;
+ALTER SCHEMA app OWNER TO app_owner;
+ALTER TABLE app.orders OWNER TO app_owner;
+ALTER SEQUENCE app.orders_id_seq OWNER TO app_owner;`,
+        inspect: `SELECT
+  (SELECT schema_owner FROM information_schema.schemata WHERE schema_name = 'app') AS schema_owner,
+  (SELECT tableowner FROM pg_tables WHERE schemaname = 'app' AND tablename = 'orders') AS object_owner,
+  pg_has_role('deploy', 'app_owner', 'SET') AS deploy_can_set;`,
+        cards: (row) => [
+          [
+            "Schema owner",
+            row?.schema_owner,
+            row?.schema_owner === "app_owner" ? "pass" : "blocked",
+          ],
+          [
+            "Table owner",
+            row?.object_owner,
+            row?.object_owner === "app_owner" ? "pass" : "blocked",
+          ],
+          [
+            "Deploy",
+            row?.deploy_can_set ? "holds app_owner" : "blocked",
+            row?.deploy_can_set ? "pass" : "blocked",
+          ],
+        ],
+        expect: (output, row) =>
+          !output.error &&
+          row?.schema_owner === "app_owner" &&
+          row?.object_owner === "app_owner" &&
+          row?.deploy_can_set,
+        observation:
+          "Ownership now belongs to a durable role. Because the membership inherits, deploy already carries the owner’s authority—and it may also become app_owner explicitly.",
+      },
+      {
+        title: "Run the migration as the owner",
+        why: "Inheritance already lets deploy pass owner checks, so retrying the plain ALTER would now succeed. The recipe still matters for a different reason: anything deploy creates belongs to deploy, not app_owner. SET ROLE makes app_owner the creator, so owner-only DDL and newly created objects both land on the durable role.",
+        prompt: "Use SET ROLE as the migration recipe, then apply the ALTER.",
+        setup: ownerSeed,
+        role: "deploy",
+        sql: `SET ROLE app_owner;
+ALTER TABLE app.orders ADD COLUMN source text;
+RESET ROLE;`,
+        inspect: `SELECT tableowner AS object_owner,
+  EXISTS (SELECT 1 FROM information_schema.columns WHERE table_schema = 'app' AND table_name = 'orders' AND column_name = 'source') AS migration_applied
+FROM pg_tables WHERE schemaname = 'app' AND tablename = 'orders';`,
+        cards: (row) => [
+          ["Deploy", "SET ROLE app_owner", "focus"],
+          [
+            "Table owner",
+            row?.object_owner,
+            row?.object_owner === "app_owner" ? "pass" : "blocked",
+          ],
+          [
+            "Migration",
+            row?.migration_applied ? "applied" : "missing",
+            row?.migration_applied ? "pass" : "blocked",
+          ],
+        ],
+        expect: (output, row) => !output.error && row?.migration_applied,
+        observation:
+          "The migration ran as the owner. Try deleting the SET ROLE line: the ALTER still succeeds through inheritance, but a table the migration created would have belonged to deploy. That mistake is exactly the next chapter.",
+      },
+    ],
+  },
+  defaults: {
+    eyebrow: "Chapter 5 · Future objects",
+    title: "A new refunds table breaks the report",
+    description:
+      "Existing wildcard grants are a snapshot. A migration created by the wrong role reveals that ownership and default privileges must line up.",
+    next: {
+      href: "/docs/postgresql-offboarding",
+      title: "Offboard the founder safely",
+    },
+    actors: [
+      ["deploy", "Deploy"],
+      ["reporting_app", "Reporting app"],
+      ["postgres", "Database admin"],
+    ],
+    steps: [
+      {
+        title: "Create refunds as deploy",
+        why: "This is the tempting migration: deploy skips the SET ROLE recipe and creates the table as itself. The object is new, so the old orders grant says nothing about it.",
+        prompt: "Create the refunds table as deploy.",
+        setup: ownerSeed,
+        role: "deploy",
+        sql: `CREATE TABLE app.refunds (
+  id bigint PRIMARY KEY,
+  order_id bigint NOT NULL,
+  total_cents bigint NOT NULL
+);
+INSERT INTO app.refunds VALUES (1, 1, 500);`,
+        inspect: privilegeInspection("reporting_app", "app.refunds"),
+        cards: (row) => [
+          [
+            "Creator / owner",
+            row?.object_owner,
+            row?.object_owner === "deploy" ? "focus" : "neutral",
+          ],
+          [
+            "Reader schema",
+            row?.schema_usage ? "USAGE" : "blocked",
+            row?.schema_usage ? "pass" : "blocked",
+          ],
+          [
+            "Reader table",
+            row?.object_select ? "SELECT" : "missing",
+            row?.object_select ? "pass" : "blocked",
+          ],
+        ],
+        expect: (output, row) =>
+          !output.error &&
+          row?.object_owner === "deploy" &&
+          !row?.object_select,
+        observation:
+          "The table belongs to deploy and orders_reader received no SELECT. Defaults are attached to the role that creates the object.",
+      },
+      {
+        title: "Observe the broken report",
+        why: "The application still reaches the app schema, but PostgreSQL evaluates the new table’s ACL independently.",
+        prompt: "Read refunds as reporting_app.",
+        setup: defaultsSeed,
+        role: "reporting_app",
+        sql: "SELECT * FROM app.refunds ORDER BY id;",
+        inspect: privilegeInspection("reporting_app", "app.refunds"),
+        cards: basicCards("Reporting app", "denied"),
+        expect: (output) =>
+          output.error?.includes("permission denied for table refunds"),
+        observation:
+          "The report fails at the object gate. Schema USAGE still works; SELECT on this new table is missing.",
+      },
+      {
+        title: "Repair the existing table",
+        why: "Two things are wrong with refunds: the reader has no SELECT, and the object belongs to a deployment login instead of the durable owner. Fix both on the existing object first.",
+        prompt: "Grant the missing SELECT and transfer refunds to app_owner.",
+        setup: defaultsSeed,
+        role: "postgres",
+        sql: `GRANT SELECT ON app.refunds TO orders_reader;
+ALTER TABLE app.refunds OWNER TO app_owner;`,
+        inspect: privilegeInspection("reporting_app", "app.refunds"),
+        cards: (row) => [
+          [
+            "Table owner",
+            row?.object_owner,
+            row?.object_owner === "app_owner" ? "pass" : "blocked",
+          ],
+          [
+            "Reader table",
+            row?.object_select ? "SELECT" : "missing",
+            row?.object_select ? "pass" : "blocked",
+          ],
+          ["Scope", "this table only", "focus"],
+        ],
+        expect: (output, row) =>
+          !output.error &&
+          row?.object_owner === "app_owner" &&
+          row?.object_select,
+        observation:
+          "refunds is repaired—but this fixed one existing table once. The next migration would break the report all over again.",
+      },
+      {
+        title: "Make tomorrow automatic",
+        why: "A default privilege changes what app_owner grants at the moment it creates a future object. It only fires when app_owner really is the creator, which is why the SET ROLE migration recipe matters.",
+        prompt:
+          "Configure the owner’s defaults, then create the next table through app_owner and watch the grant appear on its own.",
+        setup: repairedSeed,
+        role: "postgres",
+        sql: `ALTER DEFAULT PRIVILEGES FOR ROLE app_owner IN SCHEMA app
+  GRANT SELECT ON TABLES TO orders_reader;
+SET ROLE app_owner;
+CREATE TABLE app.shipments (
+  id bigint PRIMARY KEY,
+  order_id bigint NOT NULL,
+  status text NOT NULL
+);
+INSERT INTO app.shipments VALUES (1, 1, 'sent');
+RESET ROLE;`,
+        inspect: privilegeInspection("reporting_app", "app.shipments"),
+        cards: (row) => [
+          [
+            "Creating role",
+            row?.object_owner,
+            row?.object_owner === "app_owner" ? "pass" : "blocked",
+          ],
+          [
+            "Default applied",
+            row?.object_select ? "SELECT" : "missing",
+            row?.object_select ? "pass" : "blocked",
+          ],
+          [
+            "Reporting app",
+            row?.object_select ? "ready" : "denied",
+            row?.object_select ? "pass" : "blocked",
+          ],
+        ],
+        expect: (output, row) =>
+          !output.error &&
+          row?.object_owner === "app_owner" &&
+          row?.object_select,
+        observation:
+          "Nobody granted SELECT on shipments—the default fired because app_owner created the table. Wildcards repair the present; default privileges prepare the future.",
+      },
+    ],
+  },
+  offboarding: {
+    eyebrow: "Chapter 6 · Offboarding",
+    title: "Priya is leaving Acme",
+    description:
+      "Removing a login is not enough when the role owns database objects. Let PostgreSQL expose the dependency, then retire the owner deliberately.",
+    next: {
+      href: "/docs/postgresql-role-hierarchy",
+      title: "Explore membership mechanics",
+    },
+    actors: [
+      ["postgres", "Database admin"],
+      ["priya", "Priya"],
+    ],
+    steps: [
+      {
+        title: "Try the obvious DROP ROLE",
+        why: "PostgreSQL refuses to erase a role while objects still depend on it. That refusal protects the database from an incomplete offboarding.",
+        prompt: "Try to drop Priya.",
+        setup: offboardingSeed,
+        role: "postgres",
+        sql: "DROP ROLE priya;",
+        inspect: `SELECT count(*)::int AS owned_objects FROM pg_class c JOIN pg_roles r ON r.oid = c.relowner WHERE r.rolname = 'priya';`,
+        cards: (row, output) => [
+          ["Role", "Priya", "focus"],
+          [
+            "Owned objects",
+            String(row?.owned_objects ?? "?"),
+            row?.owned_objects > 0 ? "blocked" : "pass",
+          ],
+          [
+            "DROP ROLE",
+            output.error ? "refused" : "complete",
+            output.error ? "blocked" : "pass",
+          ],
+        ],
+        expect: (output, row) =>
+          output.error?.includes("cannot be dropped") && row?.owned_objects > 0,
+        observation:
+          "The failure is useful evidence: legacy.customer_notes still belongs to Priya.",
+      },
+      {
+        title: "Transfer ownership, then retire",
+        why: "The durable app_owner role is the successor created for exactly this moment. Reassign owned objects before removing Priya’s remaining grants and role.",
+        prompt: "Run the retirement sequence in dependency order.",
+        setup: offboardingSeed,
+        role: "postgres",
+        sql: `REASSIGN OWNED BY priya TO app_owner;
+DROP OWNED BY priya;
+DROP ROLE priya;`,
+        inspect: `SELECT
+  NOT EXISTS (SELECT 1 FROM pg_roles WHERE rolname = 'priya') AS role_removed,
+  (SELECT tableowner FROM pg_tables WHERE schemaname = 'legacy' AND tablename = 'customer_notes') AS successor_owner;`,
+        cards: (row) => [
+          [
+            "Priya",
+            row?.role_removed ? "role removed" : "still present",
+            row?.role_removed ? "pass" : "blocked",
+          ],
+          [
+            "Legacy objects",
+            row?.successor_owner,
+            row?.successor_owner === "app_owner" ? "pass" : "blocked",
+          ],
+          ["Application owner", "durable successor", "pass"],
+        ],
+        expect: (output, row) =>
+          !output.error &&
+          row?.role_removed &&
+          row?.successor_owner === "app_owner",
+        observation:
+          "The human login is gone and its objects survived under app_owner. Session termination is a separate operational step outside this browser lab.",
+      },
+    ],
+  },
+  mechanics: {
+    eyebrow: "Chapter 7 · Membership mechanics",
+    title: "Inside the membership edge",
+    description:
+      "Nest Bob behind an analyst job role, then take one membership edge apart: automatic inheritance, deliberate SET ROLE, and delegated administration.",
+    next: {
+      href: "/docs/postgresql-security-review",
+      title: "Audit PUBLIC, SECURITY DEFINER, and delegation",
+    },
+    actors: [
+      ["postgres", "Database admin"],
+      ["bob", "Bob"],
+      ["team_lead", "Team lead"],
+      ["dana", "Dana"],
+    ],
+    steps: [
+      {
+        title: "Nest the reporting membership",
+        why: "Acme now has several reader capabilities, so it introduces an analyst job role between the person and the capability. Membership is transitive: privileges can travel more than one edge.",
+        prompt:
+          "Insert analyst between Bob and orders_reader, and remove Bob’s direct edge.",
+        setup: mechanicsSeed,
+        role: "postgres",
+        sql: `GRANT orders_reader TO analyst;
+GRANT analyst TO bob;
+REVOKE orders_reader FROM bob;`,
+        inspect: `SELECT
+  pg_has_role('bob', 'analyst', 'USAGE') AS via_analyst,
+  pg_has_role('bob', 'orders_reader', 'USAGE') AS via_reader,
+  has_table_privilege('bob', 'app.orders', 'SELECT') AS object_select;`,
+        cards: (row) => [
+          [
+            "Edge 1",
+            row?.via_analyst ? "bob → analyst" : "missing",
+            row?.via_analyst ? "pass" : "blocked",
+          ],
+          [
+            "Edge 2",
+            row?.via_reader ? "analyst → orders_reader" : "missing",
+            row?.via_reader ? "pass" : "blocked",
+          ],
+          [
+            "Table",
+            row?.object_select ? "SELECT" : "blocked",
+            row?.object_select ? "pass" : "blocked",
+          ],
+        ],
+        expect: (output, row) =>
+          !output.error &&
+          row?.via_analyst &&
+          row?.via_reader &&
+          row?.object_select,
+        observation:
+          "Bob’s direct membership is gone, yet the table privilege still reaches him: bob → analyst → orders_reader. Object privileges stay on the leaf capability.",
+      },
+      {
+        title: "Prove it as Bob",
+        why: "The report query is the ground truth. If the nested graph is right, nothing about the query changes.",
+        prompt: "Run the report as Bob.",
+        setup: nestedSeed,
+        role: "bob",
+        sql: "SELECT * FROM app.orders ORDER BY id;",
+        inspect: privilegeInspection("bob"),
+        cards: basicCards("Bob", "2 rows"),
+        expect: (output) =>
+          !output.error && output.results[0]?.rows.length === 2,
+        observation:
+          "PostgreSQL resolved Bob’s privileges through two membership edges without any direct grant on the login.",
+      },
+      {
+        title: "Grant the edge backwards",
+        why: "GRANT takes two role names, and reversing them reverses the graph. Create the mistake instead of only reading about it.",
+        prompt:
+          "Remove the correct capability edge and grant it in the wrong direction.",
+        setup: nestedSeed,
+        role: "postgres",
+        sql: `REVOKE orders_reader FROM analyst;
+GRANT analyst TO orders_reader;`,
+        inspect: `SELECT
+  pg_has_role('bob', 'analyst', 'USAGE') AS via_analyst,
+  pg_has_role('bob', 'orders_reader', 'USAGE') AS via_reader,
+  has_table_privilege('bob', 'app.orders', 'SELECT') AS object_select;`,
+        cards: (row) => [
+          [
+            "Edge 1",
+            row?.via_analyst ? "bob → analyst" : "missing",
+            row?.via_analyst ? "pass" : "blocked",
+          ],
+          [
+            "Edge 2",
+            row?.via_reader ? "restored" : "reversed",
+            row?.via_reader ? "pass" : "blocked",
+          ],
+          [
+            "Table",
+            row?.object_select ? "SELECT" : "blocked",
+            row?.object_select ? "pass" : "blocked",
+          ],
+        ],
+        expect: (output, row) =>
+          !output.error &&
+          row?.via_analyst &&
+          !row?.via_reader &&
+          !row?.object_select,
+        observation:
+          "GRANT analyst TO orders_reader made orders_reader a member of analyst—the opposite relationship—so nothing flows toward Bob anymore. Read every edge as “member of”, from the granted role toward the member. Switch to Bob and run the report to feel the breakage.",
+      },
+      {
+        title: "Turn off automatic inheritance",
+        why: "PostgreSQL 16 and later stores three independent facts on each membership edge: MEMBER, INHERIT, and SET. Rebuild Bob’s edge so membership survives but privileges stop flowing automatically.",
+        prompt: "Recreate Bob’s analyst membership with INHERIT FALSE.",
+        setup: nestedSeed,
+        role: "postgres",
+        sql: `REVOKE analyst FROM bob;
+GRANT analyst TO bob WITH INHERIT FALSE, SET TRUE;`,
+        inspect: `SELECT
+  pg_has_role('bob', 'analyst', 'MEMBER') AS member,
+  pg_has_role('bob', 'analyst', 'USAGE') AS inherits,
+  pg_has_role('bob', 'analyst', 'SET') AS can_set;`,
+        cards: (row) => [
+          [
+            "Member",
+            row?.member ? "yes" : "no",
+            row?.member ? "pass" : "blocked",
+          ],
+          [
+            "Inherits",
+            row?.inherits ? "on" : "off",
+            row?.inherits ? "pass" : "focus",
+          ],
+          [
+            "May SET ROLE",
+            row?.can_set ? "yes" : "no",
+            row?.can_set ? "pass" : "blocked",
+          ],
+        ],
+        expect: (output, row) =>
+          !output.error && row?.member && !row?.inherits && row?.can_set,
+        observation:
+          "The edge still exists, but its ordinary privileges are dormant. Membership, automatic inheritance, and permission to SET ROLE are three separate switches.",
+      },
+      {
+        title: "Feel the dormant edge",
+        why: "With inheritance off, Bob keeps the membership but loses the automatic privilege flow. The report should now fail exactly like a missing grant.",
+        prompt: "Run the report as Bob on the INHERIT FALSE edge.",
+        setup: dormantSeed,
+        role: "bob",
+        sql: "SELECT * FROM app.orders ORDER BY id;",
+        inspect: privilegeInspection("bob"),
+        cards: basicCards("Bob", "denied"),
+        expect: (output) =>
+          output.error?.includes("permission denied for schema app"),
+        observation:
+          "Bob is still a member of analyst, but nothing flows automatically anymore. Membership alone proves nothing about effective access—in either direction.",
+      },
+      {
+        title: "Become the analyst deliberately",
+        why: "SET TRUE is the switch that still works: Bob may become analyst on purpose. SET ROLE changes the effective identity instead of copying grants onto the login.",
+        prompt: "Switch the effective role, inspect both identities, and rerun the report.",
+        setup: dormantSeed,
+        role: "bob",
+        sql: `SET ROLE analyst;
+SELECT session_user, current_user;
+SELECT * FROM app.orders ORDER BY id;`,
+        inspect: `SELECT
+  pg_has_role('bob', 'analyst', 'SET') AS can_set,
+  has_table_privilege('analyst', 'app.orders', 'SELECT') AS analyst_select;`,
+        cards: (row, output) => [
+          ["session_user", "bob", "neutral"],
+          ["current_user", "analyst", "focus"],
+          [
+            "Report",
+            output.error ? "denied" : "2 rows",
+            output.error ? "blocked" : "pass",
+          ],
+        ],
+        expect: (output) =>
+          !output.error &&
+          output.results[0]?.rows[0]?.session_user === "bob" &&
+          output.results[0]?.rows[0]?.current_user === "analyst" &&
+          output.results[1]?.rows.length === 2,
+        observation:
+          "current_user became analyst while session_user stayed bob. PostgreSQL evaluated the analyst hierarchy for the query; Bob’s own privileges never changed.",
+      },
+      {
+        title: "Delegate membership administration",
+        why: "ADMIN is the third per-edge option: it authorizes managing who belongs to the role. It grants no data access and no permission to become the role.",
+        prompt:
+          "Give the team lead an administrative edge with inheritance and SET both off.",
+        setup: nestedSeed,
+        role: "postgres",
+        sql: `GRANT analyst TO team_lead WITH ADMIN TRUE, INHERIT FALSE, SET FALSE;`,
+        inspect: `SELECT
+  membership.admin_option AS admin,
+  membership.inherit_option AS inherits,
+  membership.set_option AS can_set
+FROM pg_auth_members AS membership
+JOIN pg_roles AS granted ON granted.oid = membership.roleid
+JOIN pg_roles AS member ON member.oid = membership.member
+WHERE granted.rolname = 'analyst' AND member.rolname = 'team_lead';`,
+        cards: (row) => [
+          [
+            "Admin",
+            row?.admin ? "on" : "off",
+            row?.admin ? "focus" : "blocked",
+          ],
+          [
+            "Inherits",
+            row?.inherits ? "on" : "off",
+            row?.inherits ? "blocked" : "pass",
+          ],
+          [
+            "May SET ROLE",
+            row?.can_set ? "yes" : "no",
+            row?.can_set ? "blocked" : "pass",
+          ],
+        ],
+        expect: (output, row) =>
+          !output.error && row?.admin && !row?.inherits && !row?.can_set,
+        observation:
+          "team_lead can now manage analyst’s membership while holding none of its privileges. pgroles manages inherit and admin per edge; it does not converge PostgreSQL’s SET option.",
+      },
+      {
+        title: "Let the team lead onboard Dana",
+        why: "Dana joins the reporting team, and the team lead—not the database administrator—makes the change. Delegation works, and it quietly creates an edge the desired-state policy has never heard of.",
+        prompt: "Grant analyst to Dana as the team lead.",
+        setup: delegatedSeed,
+        role: "team_lead",
+        sql: "GRANT analyst TO dana;",
+        inspect: `SELECT
+  pg_has_role('dana', 'orders_reader', 'USAGE') AS dana_reads,
+  has_table_privilege('team_lead', 'app.orders', 'SELECT') AS lead_reads;`,
+        cards: (row) => [
+          [
+            "Dana",
+            row?.dana_reads ? "reads via hierarchy" : "no access",
+            row?.dana_reads ? "pass" : "blocked",
+          ],
+          [
+            "Team lead",
+            row?.lead_reads ? "reads orders" : "no data access",
+            row?.lead_reads ? "blocked" : "pass",
+          ],
+          ["Policy", "edge undeclared", "focus"],
+        ],
+        expect: (output, row) =>
+          !output.error && row?.dana_reads && !row?.lead_reads,
+        observation:
+          "Dana reads orders through the hierarchy, granted by someone who cannot read it themselves. Until policy declares Dana, an authoritative pgroles plan treats this edge as drift—delegation and desired state answer different questions.",
+      },
+    ],
+  },
+  security: {
+    eyebrow: "Chapter 8 · Security review",
+    title: "The auditor asks: “Who can really do this?”",
+    description:
+      "Effective access hides outside ordinary direct ACLs. Investigate three surprises: PUBLIC, SECURITY DEFINER, and delegated grant options.",
+    next: {
+      href: "/docs/postgresql-playground",
+      title: "Audit the complete Acme database",
+    },
+    actors: [
+      ["auditor", "Auditor"],
+      ["contractor", "Contractor"],
+      ["team_lead", "Team lead"],
+      ["postgres", "Database admin"],
+    ],
+    steps: [
+      {
+        title: "Surprise 1: PUBLIC can execute",
+        why: "PostgreSQL grants EXECUTE on new functions to PUBLIC by default. Every role is part of PUBLIC, even when no named function grant exists.",
+        prompt: "Call the function as contractor, who has no declared access.",
+        setup: completeSeed,
+        role: "contractor",
+        sql: "SELECT billing_api.order_total() AS total_cents;",
+        inspect: `SELECT has_function_privilege('contractor', 'billing_api.order_total()', 'EXECUTE') AS can_execute;`,
+        cards: (row, output) => [
+          ["Contractor", "no named grant", "neutral"],
+          [
+            "PUBLIC",
+            row?.can_execute ? "EXECUTE" : "revoked",
+            row?.can_execute ? "focus" : "pass",
+          ],
+          [
+            "Function",
+            output.error ? "denied" : "runs",
+            output.error ? "blocked" : "focus",
+          ],
+        ],
+        expect: (output, row) =>
+          !output.error &&
+          row?.can_execute &&
+          output.results[0]?.rows[0]?.total_cents === 14100,
+        observation:
+          "The access is real and comes from PUBLIC. Desired policy must remove the existing privilege and the owner’s default for future functions.",
+      },
+      {
+        title: "Surprise 2: SECURITY DEFINER crosses the table ACL",
+        why: "A SECURITY DEFINER function runs with its owner’s privileges. The caller can produce a result without receiving SELECT on the underlying table.",
+        prompt:
+          "Call the approved function as auditor. The diagram will inspect the table privilege separately.",
+        setup: completeSeed,
+        role: "auditor",
+        sql: "SELECT billing_api.order_total() AS total_cents;",
+        inspect: privilegeInspection("auditor"),
+        cards: (row, output) => [
+          [
+            "Auditor table ACL",
+            row?.object_select ? "SELECT" : "none",
+            row?.object_select ? "blocked" : "pass",
+          ],
+          [
+            "Definer function",
+            output.results[0]?.rows[0]?.total_cents === 14100
+              ? "returned total"
+              : "missing",
+            "focus",
+          ],
+          [
+            "Function call",
+            output.error ? "denied" : "allowed",
+            output.error ? "blocked" : "focus",
+          ],
+        ],
+        expect: (output, row) =>
+          output.results[0]?.rows[0]?.total_cents === 14100 &&
+          !output.error &&
+          !row?.object_select,
+        observation:
+          "The function returns the total even though auditor has no SELECT on orders. Function execution is another access path.",
+      },
+      {
+        title: "Prove the underlying table is still closed",
+        why: "The successful function call did not copy its owner’s privileges onto auditor. Test the table operation directly to preserve that distinction.",
+        prompt: "Run the underlying SELECT as auditor.",
+        setup: completeSeed,
+        role: "auditor",
+        sql: "SELECT * FROM app.orders;",
+        inspect: privilegeInspection("auditor"),
+        cards: basicCards("Auditor", "denied"),
+        expect: (output, row) =>
+          output.error?.includes("permission denied for schema app") &&
+          !row?.schema_usage &&
+          !row?.object_select,
+        observation:
+          "The direct read fails. SECURITY DEFINER exposes only the operation implemented by the function, with the authority and risks of its owner.",
+      },
+      {
+        title: "Surprise 3: grant options create new grantors",
+        why: "WITH GRANT OPTION lets its holder delegate an object privilege. That power can create access outside the role hierarchy. This step impersonates team_lead mid-script with SET SESSION AUTHORIZATION—the same switch the role selector performs before your SQL runs.",
+        prompt:
+          "Give team_lead a grant option, then let that role grant SELECT to contractor.",
+        setup: completeSeed,
+        role: "postgres",
+        sql: `GRANT USAGE ON SCHEMA app TO contractor;
+GRANT USAGE ON SCHEMA app TO team_lead;
+GRANT SELECT ON app.orders TO team_lead WITH GRANT OPTION;
+SET SESSION AUTHORIZATION team_lead;
+GRANT SELECT ON app.orders TO contractor;
+SET SESSION AUTHORIZATION postgres;`,
+        inspect: privilegeInspection("contractor"),
+        cards: basicCards("Contractor", "delegated access"),
+        expect: (output, row) =>
+          !output.error && row?.schema_usage && row?.object_select,
+        observation:
+          "Contractor can now read through a grant made by team_lead. pgroles does not model application grant options, so this boundary needs a separate review.",
+      },
+    ],
+  },
+  playground: {
+    eyebrow: "Final lab · The Acme database",
+    title: "Audit the whole story yourself",
+    description:
+      "The finished database is open. Choose a role, edit any query, and investigate every access path without changing the canonical scenario.",
+    actors: [
+      ["alice", "Alice"],
+      ["bob", "Bob"],
+      ["dana", "Dana"],
+      ["reporting_app", "Reporting app"],
+      ["deploy", "Deploy"],
+      ["auditor", "Auditor"],
+      ["contractor", "Contractor"],
+      ["team_lead", "Team lead"],
+      ["postgres", "Database admin"],
+    ],
+    open: true,
+    steps: [
+      {
+        title: "Find every path to orders",
+        why: "Start with evidence, not assumptions. Change the role or query freely; every run gets a fresh copy of Acme’s completed database.",
+        prompt:
+          "Which roles can read orders, call the definer function, administer a membership, or own an object?",
+        setup: completeSeed,
+        role: "bob",
+        sql: `SELECT
+  session_user,
+  current_user,
+  has_schema_privilege(current_user, 'app', 'USAGE') AS schema_usage,
+  has_table_privilege(current_user, 'app.orders', 'SELECT') AS orders_select,
+  pg_has_role(current_user, 'orders_reader', 'USAGE') AS through_reader;`,
+        inspect: privilegeInspection("bob"),
+        cards: basicCards("Selected role", "inspect output"),
+        expect: () => true,
+        observation:
+          "There is no completion test here. Use the prompts below as an open-ended security review.",
+        challenges: [
+          [
+            "Can Bob read orders?",
+            "SELECT * FROM app.orders ORDER BY id;",
+            "bob",
+          ],
+          [
+            "Did Alice’s offboarding hold?",
+            "SELECT * FROM app.orders ORDER BY id;",
+            "alice",
+          ],
+          [
+            "Who owns Priya’s legacy notes now?",
+            `SELECT schemaname, tablename, tableowner FROM pg_tables WHERE schemaname = 'legacy';`,
+            "postgres",
+          ],
+          [
+            "What does deploy own?",
+            `SELECT schemaname, tablename, tableowner FROM pg_tables WHERE tableowner = current_user;`,
+            "deploy",
+          ],
+          [
+            "Can contractor call the function?",
+            "SELECT billing_api.order_total();",
+            "contractor",
+          ],
+          [
+            "Who belongs to analyst?",
+            `SELECT member.rolname FROM pg_auth_members m JOIN pg_roles granted ON granted.oid = m.roleid JOIN pg_roles member ON member.oid = m.member WHERE granted.rolname = 'analyst';`,
+            "postgres",
+          ],
+        ],
+      },
+    ],
+  },
+};
+
+const acmeChapterSeeds = {
+  founderSeed,
+  schemaOnlySeed,
+  directAccessSeed,
+  capabilitySeed,
+  postDriftSeed,
+  ownerSeed,
+  defaultsSeed,
+  repairedSeed,
+  durableSeed,
+  offboardingSeed,
+  retiredSeed,
+  mechanicsSeed,
+  nestedSeed,
+  dormantSeed,
+  delegatedSeed,
+  completeSeed,
+};
+
+export { chapters, acmeChapterSeeds, privilegeInspection };
