+++ docs/src/components/AcmeStoryLabs.jsx
@@ -2,843 +2,7 @@ import { Fragment, useEffect, useMemo, useRef, useState } from "react";
 import Link from "next/link";
 import Highlight, { defaultProps } from "prism-react-renderer";
 
-const ordersTable = `CREATE TABLE app.orders (
-  id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-  customer text NOT NULL,
-  total_cents bigint NOT NULL
-);
-INSERT INTO app.orders (customer, total_cents)
-VALUES ('Acme', 4200), ('Globex', 9900);`;
-
-const founderSeed = `
-CREATE ROLE priya LOGIN;
-CREATE ROLE alice LOGIN;
-CREATE ROLE bob LOGIN;
-CREATE ROLE reporting_app LOGIN;
-CREATE ROLE deploy LOGIN;
-GRANT CREATE ON DATABASE postgres TO priya;
-SET ROLE priya;
-CREATE SCHEMA app AUTHORIZATION priya;
-${ordersTable}
-RESET ROLE;`;
-
-const directAccessSeed = `${founderSeed}
-GRANT USAGE ON SCHEMA app TO alice;
-GRANT SELECT ON app.orders TO alice;`;
-
-const capabilitySeed = `${directAccessSeed}
-CREATE ROLE orders_reader;
-GRANT USAGE ON SCHEMA app TO orders_reader;
-GRANT SELECT ON app.orders TO orders_reader;
-GRANT orders_reader TO alice, reporting_app;`;
-
-const postDriftSeed = `${capabilitySeed}
-GRANT orders_reader TO bob;
-REVOKE orders_reader FROM alice;
-REVOKE SELECT ON app.orders FROM alice;
-REVOKE USAGE ON SCHEMA app FROM alice;`;
-
-const ownerSeed = `${postDriftSeed}
-CREATE ROLE app_owner;
-GRANT app_owner TO deploy;
-ALTER SCHEMA app OWNER TO app_owner;
-ALTER TABLE app.orders OWNER TO app_owner;
-ALTER SEQUENCE app.orders_id_seq OWNER TO app_owner;`;
-
-const defaultsSeed = `${ownerSeed}
-SET SESSION AUTHORIZATION deploy;
-CREATE TABLE app.refunds (
-  id bigint PRIMARY KEY,
-  order_id bigint NOT NULL,
-  total_cents bigint NOT NULL
-);
-INSERT INTO app.refunds VALUES (1, 1, 500);
-SET SESSION AUTHORIZATION postgres;`;
-
-const durableSeed = `${ownerSeed}
-SET ROLE app_owner;
-ALTER DEFAULT PRIVILEGES FOR ROLE app_owner IN SCHEMA app
-  GRANT SELECT ON TABLES TO orders_reader;
-CREATE TABLE app.refunds (
-  id bigint PRIMARY KEY,
-  order_id bigint NOT NULL,
-  total_cents bigint NOT NULL
-);
-INSERT INTO app.refunds VALUES (1, 1, 500);
-GRANT SELECT ON app.refunds TO orders_reader;
-CREATE TABLE app.shipments (
-  id bigint PRIMARY KEY,
-  order_id bigint NOT NULL,
-  status text NOT NULL
-);
-INSERT INTO app.shipments VALUES (1, 1, 'sent');
-RESET ROLE;`;
-
-const offboardingSeed = `${durableSeed}
-GRANT CREATE ON DATABASE postgres TO priya;
-SET ROLE priya;
-CREATE SCHEMA legacy AUTHORIZATION priya;
-CREATE TABLE legacy.customer_notes (id bigint PRIMARY KEY, note text);
-RESET ROLE;`;
-
-const completeSeed = `${durableSeed}
-DROP OWNED BY priya;
-DROP ROLE priya;
-CREATE ROLE team_lead LOGIN;
-CREATE ROLE contractor LOGIN;
-CREATE ROLE auditor LOGIN;
-CREATE SCHEMA billing_api AUTHORIZATION app_owner;
-GRANT USAGE ON SCHEMA billing_api TO PUBLIC;
-SET ROLE app_owner;
-CREATE FUNCTION billing_api.order_total()
-RETURNS bigint LANGUAGE sql SECURITY DEFINER
-SET search_path = pg_catalog, app
-AS 'SELECT sum(total_cents)::bigint FROM app.orders';
-RESET ROLE;
-GRANT EXECUTE ON FUNCTION billing_api.order_total() TO auditor;`;
-
-function privilegeInspection(role, object = "app.orders") {
-  return `SELECT
-  has_schema_privilege('${role}', 'app', 'USAGE') AS schema_usage,
-  has_table_privilege('${role}', '${object}', 'SELECT') AS object_select,
-  CASE
-    WHEN EXISTS (SELECT 1 FROM pg_roles WHERE rolname = 'orders_reader')
-    THEN pg_has_role('${role}', 'orders_reader', 'USAGE')
-    ELSE false
-  END AS inherits_reader,
-  (SELECT tableowner FROM pg_tables
-    WHERE schemaname = 'app' AND tablename = '${
-      object.split(".")[1]
-    }') AS object_owner;`;
-}
-
-const basicCards = (role, resultLabel) => (row, output) =>
-  [
-    ["Role", role, "neutral"],
-    [
-      "Membership",
-      row?.inherits_reader ? "orders_reader" : "none",
-      row?.inherits_reader ? "pass" : "neutral",
-    ],
-    [
-      "Schema",
-      row?.schema_usage ? "USAGE" : "blocked",
-      row?.schema_usage ? "pass" : "blocked",
-    ],
-    [
-      "Table",
-      row?.object_select ? "SELECT" : "blocked",
-      row?.object_select ? "pass" : "blocked",
-    ],
-    [
-      "Result",
-      output.error ? "denied" : resultLabel,
-      output.error ? "blocked" : "pass",
-    ],
-  ];
-
-const chapters = {
-  gates: {
-    eyebrow: "Chapter 1 · The permission chain",
-    title: "Alice gets the first report working",
-    description:
-      "Acme needs an orders report today. Give Alice the smallest working path, then prove it with the application query.",
-    next: {
-      href: "/docs/postgresql-capability-roles",
-      title: "Replace direct grants with a capability role",
-    },
-    actors: [
-      ["postgres", "Database admin"],
-      ["alice", "Alice"],
-    ],
-    steps: [
-      {
-        title: "Grant both gates",
-        why: "PostgreSQL checks two separate gates: USAGE on app, then SELECT on orders.",
-        prompt: "Grant Alice both parts of that path.",
-        setup: founderSeed,
-        role: "postgres",
-        sql: `GRANT USAGE ON SCHEMA app TO alice;
-GRANT SELECT ON app.orders TO alice;`,
-        inspect: privilegeInspection("alice"),
-        cards: basicCards("Alice", "path ready"),
-        expect: (output, row) =>
-          !output.error && row?.schema_usage && row?.object_select,
-        observation:
-          "Alice now has a complete path: PostgreSQL can resolve app.orders, then authorize SELECT.",
-      },
-      {
-        title: "Run the real report",
-        why: "Permissions matter only when the application operation succeeds. Run the same SELECT the report will send.",
-        prompt: "Query orders as Alice.",
-        setup: directAccessSeed,
-        role: "alice",
-        sql: "SELECT * FROM app.orders ORDER BY id;",
-        inspect: privilegeInspection("alice"),
-        cards: basicCards("Alice", "2 rows"),
-        expect: (output) =>
-          !output.error && output.results[0]?.rows.length === 2,
-        observation:
-          "The report works. Remember that these convenient direct grants will survive until someone explicitly removes them.",
-      },
-    ],
-  },
-  capabilities: {
-    eyebrow: "Chapter 2 · Capability roles",
-    title: "A reporting application joins Alice",
-    description:
-      "A second consumer needs the same access. Name the capability once instead of copying grants onto every login.",
-    next: {
-      href: "/docs/postgresql-access-drift",
-      title: "Make the old direct grants resurface",
-    },
-    actors: [
-      ["postgres", "Database admin"],
-      ["alice", "Alice"],
-      ["reporting_app", "Reporting app"],
-    ],
-    steps: [
-      {
-        title: "Create orders_reader",
-        why: "The permission bundle is a job, not a person. A NOLOGIN role gives that job a durable name and keeps object ACLs independent from staff changes.",
-        prompt:
-          "Move the reusable grants onto orders_reader and make both consumers members. Deliberately leave Alice’s Chapter 1 grants in place.",
-        setup: directAccessSeed,
-        role: "postgres",
-        sql: `CREATE ROLE orders_reader;
-GRANT USAGE ON SCHEMA app TO orders_reader;
-GRANT SELECT ON app.orders TO orders_reader;
-GRANT orders_reader TO alice, reporting_app;`,
-        inspect: privilegeInspection("reporting_app"),
-        cards: basicCards("Reporting app", "path ready"),
-        expect: (output, row) =>
-          !output.error &&
-          row?.inherits_reader &&
-          row?.schema_usage &&
-          row?.object_select,
-        observation:
-          "One membership now gives the reporting application the same complete path as Alice.",
-      },
-      {
-        title: "Prove the application role works",
-        why: "A policy graph is useful only if the actual login inherits the intended privileges.",
-        prompt: "Run the report as reporting_app.",
-        setup: capabilitySeed,
-        role: "reporting_app",
-        sql: "SELECT * FROM app.orders ORDER BY id;",
-        inspect: privilegeInspection("reporting_app"),
-        cards: basicCards("Reporting app", "2 rows"),
-        expect: (output) =>
-          !output.error && output.results[0]?.rows.length === 2,
-        observation:
-          "The login has no direct table grant. Its membership in orders_reader supplies both gates.",
-      },
-      {
-        title: "Count Alice’s access paths",
-        why: "Adding the cleaner path did not remove the old one. Alice can now read through membership and through the direct grants you created in Chapter 1.",
-        prompt: "Ask PostgreSQL which paths still exist.",
-        setup: capabilitySeed,
-        role: "postgres",
-        sql: `SELECT
-  pg_has_role('alice', 'orders_reader', 'USAGE') AS through_membership,
-  EXISTS (
-    SELECT 1 FROM information_schema.table_privileges
-    WHERE grantee = 'alice' AND table_schema = 'app'
-      AND table_name = 'orders' AND privilege_type = 'SELECT'
-  ) AS direct_table_grant;`,
-        inspect: privilegeInspection("alice"),
-        cards: (row) => [
-          [
-            "Alice path 1",
-            row?.inherits_reader ? "membership" : "missing",
-            row?.inherits_reader ? "pass" : "blocked",
-          ],
-          [
-            "Alice path 2",
-            row?.object_select ? "direct ACL remains" : "missing",
-            row?.object_select ? "focus" : "blocked",
-          ],
-          ["Lesson", "two paths", "focus"],
-        ],
-        expect: (output) =>
-          output.results[0]?.rows[0]?.through_membership === true &&
-          output.results[0]?.rows[0]?.direct_table_grant === true,
-        observation:
-          "This duplicate path is harmless today, but it is already authorization drift waiting for a team change.",
-      },
-    ],
-  },
-  drift: {
-    eyebrow: "Chapter 3 · Drift",
-    title: "Bob joins the team; Alice leaves it",
-    description:
-      "Membership makes the team change look simple—until the direct grants you created earlier defeat the offboarding test.",
-    next: {
-      href: "/docs/postgresql-ownership",
-      title: "Meet PostgreSQL ownership",
-    },
-    actors: [
-      ["postgres", "Database admin"],
-      ["alice", "Alice"],
-      ["bob", "Bob"],
-    ],
-    steps: [
-      {
-        title: "Change the team",
-        why: "The intended rule is now “Bob is an orders reader; Alice is not.” Membership expresses that change in two edges.",
-        prompt: "Add Bob and remove Alice from orders_reader.",
-        setup: capabilitySeed,
-        role: "postgres",
-        sql: `GRANT orders_reader TO bob;
-REVOKE orders_reader FROM alice;`,
-        inspect: privilegeInspection("alice"),
-        cards: (row) => [
-          [
-            "Alice membership",
-            row?.inherits_reader ? "still present" : "removed",
-            row?.inherits_reader ? "blocked" : "pass",
-          ],
-          [
-            "Schema",
-            row?.schema_usage ? "still reachable" : "blocked",
-            row?.schema_usage ? "focus" : "pass",
-          ],
-          [
-            "Table",
-            row?.object_select ? "still readable" : "blocked",
-            row?.object_select ? "focus" : "pass",
-          ],
-        ],
-        expect: (output, row) =>
-          !output.error &&
-          !row?.inherits_reader &&
-          row?.schema_usage &&
-          row?.object_select,
-        observation:
-          "The membership is gone, but Alice’s effective access is not. PostgreSQL can still see the direct Chapter 1 ACLs.",
-      },
-      {
-        title: "Make the failed offboarding visible",
-        why: "Never infer loss of access from one revoked edge. Test the same operation the role used before.",
-        prompt: "Run the report as Alice after her membership was removed.",
-        setup: `${capabilitySeed}\nGRANT orders_reader TO bob;\nREVOKE orders_reader FROM alice;`,
-        role: "alice",
-        sql: "SELECT * FROM app.orders ORDER BY id;",
-        inspect: privilegeInspection("alice"),
-        cards: basicCards("Alice", "still 2 rows"),
-        expect: (output) =>
-          !output.error && output.results[0]?.rows.length === 2,
-        observation:
-          "The query still succeeds. Effective access is the union of every surviving path, not the edge you meant to remove.",
-      },
-      {
-        title: "Remove the undeclared path",
-        why: "The direct schema and table ACLs no longer belong in desired policy. Revoking both makes membership the only path.",
-        prompt: "Remove Alice’s old direct grants.",
-        setup: `${capabilitySeed}\nGRANT orders_reader TO bob;\nREVOKE orders_reader FROM alice;`,
-        role: "postgres",
-        sql: `REVOKE SELECT ON app.orders FROM alice;
-REVOKE USAGE ON SCHEMA app FROM alice;`,
-        inspect: privilegeInspection("alice"),
-        cards: basicCards("Alice", "denied"),
-        expect: (output, row) =>
-          !output.error &&
-          !row?.inherits_reader &&
-          !row?.schema_usage &&
-          !row?.object_select,
-        observation:
-          "Alice now has no surviving path. An authoritative pgroles plan would have exposed these undeclared grants before the team change.",
-      },
-      {
-        title: "Verify the negative case",
-        why: "Offboarding is complete only when the old operation fails and the replacement reader still succeeds.",
-        prompt: "Try the report as Alice. Then switch to Bob and run it again.",
-        setup: postDriftSeed,
-        role: "alice",
-        sql: "SELECT * FROM app.orders ORDER BY id;",
-        inspect: privilegeInspection("alice"),
-        cards: basicCards("Alice", "denied"),
-        expect: (output) =>
-          output.error?.includes("permission denied for schema app"),
-        observation:
-          "The denial is the evidence. Use the role selector to confirm Bob receives the same rows Alice used to receive.",
-      },
-    ],
-  },
-  ownership: {
-    eyebrow: "Chapter 4 · Ownership",
-    title: "The migration runner hits Priya’s old table",
-    description:
-      "The original founder created orders herself. Object grants can allow data access, but only an owner—or a role acting as it—can change the table.",
-    next: {
-      href: "/docs/postgresql-default-privileges",
-      title: "Keep future tables readable",
-    },
-    actors: [
-      ["deploy", "Deploy"],
-      ["postgres", "Database admin"],
-      ["app_owner", "App owner"],
-    ],
-    steps: [
-      {
-        title: "Let deploy try the migration",
-        why: "Deploy can connect, but Acme never gave it ownership of Priya’s table. ALTER TABLE checks ownership rather than an ordinary ACL privilege.",
-        prompt: "Run the migration as deploy.",
-        setup: `${postDriftSeed}
-GRANT USAGE ON SCHEMA app TO deploy;`,
-        role: "deploy",
-        sql: "ALTER TABLE app.orders ADD COLUMN source text;",
-        inspect: `SELECT tableowner AS object_owner FROM pg_tables WHERE schemaname = 'app' AND tablename = 'orders';`,
-        cards: (row, output) => [
-          ["Current role", "Deploy", "neutral"],
-          ["Table owner", row?.object_owner || "unknown", "focus"],
-          [
-            "ALTER TABLE",
-            output.error ? "owner required" : "changed",
-            output.error ? "blocked" : "pass",
-          ],
-        ],
-        expect: (output) =>
-          output.error?.includes("must be owner of table orders"),
-        observation:
-          "No combination of SELECT, INSERT, UPDATE, or DELETE grants makes deploy the owner. Ownership is a separate authority.",
-      },
-      {
-        title: "Replace the human owner",
-        why: "Production objects should outlive people. app_owner is a durable NOLOGIN role; deploy may become it for migrations without owning objects personally.",
-        prompt:
-          "Create the owner role, let deploy use it, and transfer the schema and existing objects.",
-        setup: postDriftSeed,
-        role: "postgres",
-        sql: `CREATE ROLE app_owner;
-GRANT app_owner TO deploy;
-ALTER SCHEMA app OWNER TO app_owner;
-ALTER TABLE app.orders OWNER TO app_owner;
-ALTER SEQUENCE app.orders_id_seq OWNER TO app_owner;`,
-        inspect: `SELECT
-  (SELECT schema_owner FROM information_schema.schemata WHERE schema_name = 'app') AS schema_owner,
-  (SELECT tableowner FROM pg_tables WHERE schemaname = 'app' AND tablename = 'orders') AS object_owner,
-  pg_has_role('deploy', 'app_owner', 'SET') AS deploy_can_set;`,
-        cards: (row) => [
-          [
-            "Schema owner",
-            row?.schema_owner,
-            row?.schema_owner === "app_owner" ? "pass" : "blocked",
-          ],
-          [
-            "Table owner",
-            row?.object_owner,
-            row?.object_owner === "app_owner" ? "pass" : "blocked",
-          ],
-          [
-            "Deploy",
-            row?.deploy_can_set ? "may SET ROLE" : "blocked",
-            row?.deploy_can_set ? "pass" : "blocked",
-          ],
-        ],
-        expect: (output, row) =>
-          !output.error &&
-          row?.schema_owner === "app_owner" &&
-          row?.object_owner === "app_owner" &&
-          row?.deploy_can_set,
-        observation:
-          "Ownership now belongs to a durable role. Deploy holds permission to act as that role, not permanent personal ownership.",
-      },
-      {
-        title: "Run the migration as the owner",
-        why: "Membership alone leaves deploy’s current identity unchanged. The migration deliberately becomes app_owner for owner-only DDL.",
-        prompt: "Use SET ROLE as a migration recipe, then retry ALTER TABLE.",
-        setup: ownerSeed,
-        role: "deploy",
-        sql: `SET ROLE app_owner;
-ALTER TABLE app.orders ADD COLUMN source text;
-RESET ROLE;`,
-        inspect: `SELECT tableowner AS object_owner,
-  EXISTS (SELECT 1 FROM information_schema.columns WHERE table_schema = 'app' AND table_name = 'orders' AND column_name = 'source') AS migration_applied
-FROM pg_tables WHERE schemaname = 'app' AND tablename = 'orders';`,
-        cards: (row) => [
-          ["Deploy", "SET ROLE app_owner", "focus"],
-          [
-            "Table owner",
-            row?.object_owner,
-            row?.object_owner === "app_owner" ? "pass" : "blocked",
-          ],
-          [
-            "Migration",
-            row?.migration_applied ? "applied" : "missing",
-            row?.migration_applied ? "pass" : "blocked",
-          ],
-        ],
-        expect: (output, row) => !output.error && row?.migration_applied,
-        observation:
-          "The same migration succeeds because current_user became the table owner for the duration of the DDL.",
-      },
-    ],
-  },
-  defaults: {
-    eyebrow: "Chapter 5 · Future objects",
-    title: "A new refunds table breaks the report",
-    description:
-      "Existing wildcard grants are a snapshot. A migration created by the wrong role reveals that ownership and default privileges must line up.",
-    next: {
-      href: "/docs/postgresql-offboarding",
-      title: "Offboard the founder safely",
-    },
-    actors: [
-      ["deploy", "Deploy"],
-      ["reporting_app", "Reporting app"],
-      ["postgres", "Database admin"],
-    ],
-    steps: [
-      {
-        title: "Create refunds as deploy",
-        why: "This is the tempting migration: deploy creates the table without becoming app_owner. The object is new, so the old orders grant says nothing about it.",
-        prompt: "Create the refunds table as deploy.",
-        setup: ownerSeed,
-        role: "deploy",
-        sql: `CREATE TABLE app.refunds (
-  id bigint PRIMARY KEY,
-  order_id bigint NOT NULL,
-  total_cents bigint NOT NULL
-);
-INSERT INTO app.refunds VALUES (1, 1, 500);`,
-        inspect: privilegeInspection("reporting_app", "app.refunds"),
-        cards: (row) => [
-          [
-            "Creator / owner",
-            row?.object_owner,
-            row?.object_owner === "deploy" ? "focus" : "neutral",
-          ],
-          [
-            "Reader schema",
-            row?.schema_usage ? "USAGE" : "blocked",
-            row?.schema_usage ? "pass" : "blocked",
-          ],
-          [
-            "Reader table",
-            row?.object_select ? "SELECT" : "missing",
-            row?.object_select ? "pass" : "blocked",
-          ],
-        ],
-        expect: (output, row) =>
-          !output.error &&
-          row?.object_owner === "deploy" &&
-          !row?.object_select,
-        observation:
-          "The table belongs to deploy and orders_reader received no SELECT. Defaults are attached to the role that creates the object.",
-      },
-      {
-        title: "Observe the broken report",
-        why: "The application still reaches the app schema, but PostgreSQL evaluates the new table’s ACL independently.",
-        prompt: "Read refunds as reporting_app.",
-        setup: defaultsSeed,
-        role: "reporting_app",
-        sql: "SELECT * FROM app.refunds ORDER BY id;",
-        inspect: privilegeInspection("reporting_app", "app.refunds"),
-        cards: basicCards("Reporting app", "denied"),
-        expect: (output) =>
-          output.error?.includes("permission denied for table refunds"),
-        observation:
-          "The report fails at the object gate. Schema USAGE still works; SELECT on this new table is missing.",
-      },
-      {
-        title: "Repair today and make tomorrow automatic",
-        why: "Grant the existing refunds table once, then set defaults for app_owner and ensure future migrations create objects as that owner.",
-        prompt:
-          "Repair refunds, configure the creating role’s defaults, and create shipments through the durable owner.",
-        setup: defaultsSeed,
-        role: "postgres",
-        sql: `GRANT SELECT ON app.refunds TO orders_reader;
-ALTER TABLE app.refunds OWNER TO app_owner;
-ALTER DEFAULT PRIVILEGES FOR ROLE app_owner IN SCHEMA app
-  GRANT SELECT ON TABLES TO orders_reader;
-SET ROLE app_owner;
-CREATE TABLE app.shipments (
-  id bigint PRIMARY KEY,
-  order_id bigint NOT NULL,
-  status text NOT NULL
-);
-INSERT INTO app.shipments VALUES (1, 1, 'sent');
-RESET ROLE;`,
-        inspect: privilegeInspection("reporting_app", "app.shipments"),
-        cards: (row) => [
-          [
-            "Creating role",
-            row?.object_owner,
-            row?.object_owner === "app_owner" ? "pass" : "blocked",
-          ],
-          [
-            "Default applied",
-            row?.object_select ? "SELECT" : "missing",
-            row?.object_select ? "pass" : "blocked",
-          ],
-          [
-            "Reporting app",
-            row?.object_select ? "ready" : "denied",
-            row?.object_select ? "pass" : "blocked",
-          ],
-        ],
-        expect: (output, row) =>
-          !output.error &&
-          row?.object_owner === "app_owner" &&
-          row?.object_select,
-        observation:
-          "Existing objects needed a repair grant. Future app_owner tables now receive SELECT automatically.",
-      },
-    ],
-  },
-  offboarding: {
-    eyebrow: "Chapter 6 · Offboarding",
-    title: "Priya is leaving Acme",
-    description:
-      "Removing a login is not enough when the role owns database objects. Let PostgreSQL expose the dependency, then retire the owner deliberately.",
-    next: {
-      href: "/docs/postgresql-role-hierarchy",
-      title: "Explore membership mechanics",
-    },
-    actors: [
-      ["postgres", "Database admin"],
-      ["priya", "Priya"],
-    ],
-    steps: [
-      {
-        title: "Try the obvious DROP ROLE",
-        why: "PostgreSQL refuses to erase a role while objects still depend on it. That refusal protects the database from an incomplete offboarding.",
-        prompt: "Try to drop Priya.",
-        setup: offboardingSeed,
-        role: "postgres",
-        sql: "DROP ROLE priya;",
-        inspect: `SELECT count(*)::int AS owned_objects FROM pg_class c JOIN pg_roles r ON r.oid = c.relowner WHERE r.rolname = 'priya';`,
-        cards: (row, output) => [
-          ["Role", "Priya", "focus"],
-          [
-            "Owned objects",
-            String(row?.owned_objects ?? "?"),
-            row?.owned_objects > 0 ? "blocked" : "pass",
-          ],
-          [
-            "DROP ROLE",
-            output.error ? "refused" : "complete",
-            output.error ? "blocked" : "pass",
-          ],
-        ],
-        expect: (output, row) =>
-          output.error?.includes("cannot be dropped") && row?.owned_objects > 0,
-        observation:
-          "The failure is useful evidence: legacy.customer_notes still belongs to Priya.",
-      },
-      {
-        title: "Transfer ownership, then retire",
-        why: "The durable app_owner role is the successor created for exactly this moment. Reassign owned objects before removing Priya’s remaining grants and role.",
-        prompt: "Run the retirement sequence in dependency order.",
-        setup: offboardingSeed,
-        role: "postgres",
-        sql: `REASSIGN OWNED BY priya TO app_owner;
-DROP OWNED BY priya;
-DROP ROLE priya;`,
-        inspect: `SELECT
-  NOT EXISTS (SELECT 1 FROM pg_roles WHERE rolname = 'priya') AS role_removed,
-  (SELECT tableowner FROM pg_tables WHERE schemaname = 'legacy' AND tablename = 'customer_notes') AS successor_owner;`,
-        cards: (row) => [
-          [
-            "Priya",
-            row?.role_removed ? "role removed" : "still present",
-            row?.role_removed ? "pass" : "blocked",
-          ],
-          [
-            "Legacy objects",
-            row?.successor_owner,
-            row?.successor_owner === "app_owner" ? "pass" : "blocked",
-          ],
-          ["Application owner", "durable successor", "pass"],
-        ],
-        expect: (output, row) =>
-          !output.error &&
-          row?.role_removed &&
-          row?.successor_owner === "app_owner",
-        observation:
-          "The human login is gone and its objects survived under app_owner. Session termination is a separate operational step outside this browser lab.",
-      },
-    ],
-  },
-  security: {
-    eyebrow: "Chapter 8 · Security review",
-    title: "The auditor asks: “Who can really do this?”",
-    description:
-      "Effective access hides outside ordinary direct ACLs. Investigate three surprises: PUBLIC, SECURITY DEFINER, and delegated grant options.",
-    next: {
-      href: "/docs/postgresql-playground",
-      title: "Audit the complete Acme database",
-    },
-    actors: [
-      ["auditor", "Auditor"],
-      ["contractor", "Contractor"],
-      ["team_lead", "Team lead"],
-      ["postgres", "Database admin"],
-    ],
-    steps: [
-      {
-        title: "Surprise 1: PUBLIC can execute",
-        why: "PostgreSQL grants EXECUTE on new functions to PUBLIC by default. Every role is part of PUBLIC, even when no named function grant exists.",
-        prompt: "Call the function as contractor, who has no declared access.",
-        setup: completeSeed,
-        role: "contractor",
-        sql: "SELECT billing_api.order_total() AS total_cents;",
-        inspect: `SELECT has_function_privilege('contractor', 'billing_api.order_total()', 'EXECUTE') AS can_execute;`,
-        cards: (row, output) => [
-          ["Contractor", "no named grant", "neutral"],
-          [
-            "PUBLIC",
-            row?.can_execute ? "EXECUTE" : "revoked",
-            row?.can_execute ? "focus" : "pass",
-          ],
-          [
-            "Function",
-            output.error ? "denied" : "runs",
-            output.error ? "blocked" : "focus",
-          ],
-        ],
-        expect: (output, row) =>
-          !output.error &&
-          row?.can_execute &&
-          output.results[0]?.rows[0]?.total_cents === 14100,
-        observation:
-          "The access is real and comes from PUBLIC. Desired policy must remove the existing privilege and the owner’s default for future functions.",
-      },
-      {
-        title: "Surprise 2: SECURITY DEFINER crosses the table ACL",
-        why: "A SECURITY DEFINER function runs with its owner’s privileges. The caller can produce a result without receiving SELECT on the underlying table.",
-        prompt:
-          "Call the approved function as auditor. The diagram will inspect the table privilege separately.",
-        setup: completeSeed,
-        role: "auditor",
-        sql: "SELECT billing_api.order_total() AS total_cents;",
-        inspect: privilegeInspection("auditor"),
-        cards: (row, output) => [
-          [
-            "Auditor table ACL",
-            row?.object_select ? "SELECT" : "none",
-            row?.object_select ? "blocked" : "pass",
-          ],
-          [
-            "Definer function",
-            output.results[0]?.rows[0]?.total_cents === 14100
-              ? "returned total"
-              : "missing",
-            "focus",
-          ],
-          [
-            "Function call",
-            output.error ? "denied" : "allowed",
-            output.error ? "blocked" : "focus",
-          ],
-        ],
-        expect: (output, row) =>
-          output.results[0]?.rows[0]?.total_cents === 14100 &&
-          !output.error &&
-          !row?.object_select,
-        observation:
-          "The function returns the total even though auditor has no SELECT on orders. Function execution is another access path.",
-      },
-      {
-        title: "Prove the underlying table is still closed",
-        why: "The successful function call did not copy its owner’s privileges onto auditor. Test the table operation directly to preserve that distinction.",
-        prompt: "Run the underlying SELECT as auditor.",
-        setup: completeSeed,
-        role: "auditor",
-        sql: "SELECT * FROM app.orders;",
-        inspect: privilegeInspection("auditor"),
-        cards: basicCards("Auditor", "denied"),
-        expect: (output, row) =>
-          output.error?.includes("permission denied for schema app") &&
-          !row?.schema_usage &&
-          !row?.object_select,
-        observation:
-          "The direct read fails. SECURITY DEFINER exposes only the operation implemented by the function, with the authority and risks of its owner.",
-      },
-      {
-        title: "Surprise 3: grant options create new grantors",
-        why: "WITH GRANT OPTION lets its holder delegate an object privilege. That power can create access outside the role hierarchy.",
-        prompt:
-          "Give team_lead a grant option, then let that role grant SELECT to contractor.",
-        setup: completeSeed,
-        role: "postgres",
-        sql: `GRANT USAGE ON SCHEMA app TO contractor;
-GRANT USAGE ON SCHEMA app TO team_lead;
-GRANT SELECT ON app.orders TO team_lead WITH GRANT OPTION;
-SET SESSION AUTHORIZATION team_lead;
-GRANT SELECT ON app.orders TO contractor;
-SET SESSION AUTHORIZATION postgres;`,
-        inspect: privilegeInspection("contractor"),
-        cards: basicCards("Contractor", "delegated access"),
-        expect: (output, row) =>
-          !output.error && row?.schema_usage && row?.object_select,
-        observation:
-          "Contractor can now read through a grant made by team_lead. pgroles does not model application grant options, so this boundary needs a separate review.",
-      },
-    ],
-  },
-  playground: {
-    eyebrow: "Final lab · The Acme database",
-    title: "Audit the whole story yourself",
-    description:
-      "The finished database is open. Choose a role, edit any query, and investigate every access path without changing the canonical scenario.",
-    actors: [
-      ["alice", "Alice"],
-      ["bob", "Bob"],
-      ["reporting_app", "Reporting app"],
-      ["deploy", "Deploy"],
-      ["auditor", "Auditor"],
-      ["contractor", "Contractor"],
-      ["team_lead", "Team lead"],
-      ["postgres", "Database admin"],
-    ],
-    open: true,
-    steps: [
-      {
-        title: "Find every path to orders",
-        why: "Start with evidence, not assumptions. Change the role or query freely; every run gets a fresh copy of Acme’s completed database.",
-        prompt:
-          "Which roles can read orders, call the definer function, administer a membership, or own an object?",
-        setup: completeSeed,
-        role: "alice",
-        sql: `SELECT
-  session_user,
-  current_user,
-  has_schema_privilege(current_user, 'app', 'USAGE') AS schema_usage,
-  has_table_privilege(current_user, 'app.orders', 'SELECT') AS orders_select,
-  pg_has_role(current_user, 'orders_reader', 'USAGE') AS through_reader;`,
-        inspect: privilegeInspection("alice"),
-        cards: basicCards("Selected role", "inspect output"),
-        expect: () => true,
-        observation:
-          "There is no completion test here. Use the prompts below as an open-ended security review.",
-        challenges: [
-          [
-            "Can Bob read orders?",
-            "SELECT * FROM app.orders ORDER BY id;",
-            "bob",
-          ],
-          [
-            "What does deploy own?",
-            `SELECT schemaname, tablename, tableowner FROM pg_tables WHERE tableowner = current_user;`,
-            "deploy",
-          ],
-          [
-            "Can contractor call the function?",
-            "SELECT billing_api.order_total();",
-            "contractor",
-          ],
-          [
-            "Who belongs to orders_reader?",
-            `SELECT member.rolname FROM pg_auth_members m JOIN pg_roles granted ON granted.oid = m.roleid JOIN pg_roles member ON member.oid = m.member WHERE granted.rolname = 'orders_reader';`,
-            "postgres",
-          ],
-        ],
-      },
-    ],
-  },
-};
+import { chapters } from "./AcmeStoryData.mjs";
 
 const statusStyles = {
   pass: "border-emerald-400 bg-emerald-50 dark:border-emerald-700 dark:bg-emerald-950/25",
@@ -1192,7 +356,7 @@ function AcmeStoryLab({ chapter }) {
             Step {index + 1} of {lesson.steps.length}
           </p>
           <span className="hidden font-mono text-[10px] text-stone-500 sm:inline">
-            Every run starts from this step’s canonical state
+            Every step seeds its own database — jump anywhere
           </span>
         </div>
         <ol className="mt-3 flex gap-1.5" aria-label="Lesson progress">
@@ -1201,10 +365,7 @@ function AcmeStoryLab({ chapter }) {
               <button
                 type="button"
                 onClick={() => go(itemIndex)}
-                disabled={
-                  running ||
-                  (!lesson.open && itemIndex > 0 && !completed[itemIndex - 1])
-                }
+                disabled={running}
                 className={`block h-2 w-full rounded-full ${
                   completed[itemIndex]
                     ? "bg-emerald-500"
@@ -1335,14 +496,15 @@ function AcmeStoryLab({ chapter }) {
         )}
         {output && !output.passed && canonical && (
           <p className="m-0 rounded-xl border border-rose-200 bg-rose-50 p-4 text-sm text-rose-800 dark:border-rose-900 dark:bg-rose-950/30 dark:text-rose-200">
-            This result did not match the chapter’s canonical PostgreSQL
-            outcome. Restore the example and retry.
+            This result did not match the step’s canonical PostgreSQL outcome.
+            Restore the example to compare — or continue on; every step seeds
+            its own database.
           </p>
         )}
         {output && !canonical && (
           <p className="m-0 rounded-xl border border-stone-200 bg-stone-50 p-4 text-sm text-stone-600 dark:border-stone-700 dark:bg-stone-950/30 dark:text-stone-300">
             That is PostgreSQL’s real response to your SQL. Restore the example
-            when you want to continue the guided story.
+            when you want to see the guided story’s outcome.
           </p>
         )}
         <div className="flex flex-col-reverse gap-3 border-t border-stone-200 pt-5 dark:border-stone-800 sm:flex-row sm:items-center sm:justify-between">
@@ -1358,12 +520,12 @@ function AcmeStoryLab({ chapter }) {
             <button
               type="button"
               onClick={() => go(index + 1)}
-              disabled={running || (!lesson.open && !output?.passed)}
+              disabled={running}
               className="rounded-lg border border-stone-900 bg-stone-900 px-4 py-2 text-sm font-semibold text-white disabled:cursor-not-allowed disabled:border-stone-200 disabled:bg-stone-100 disabled:text-stone-400 dark:border-white dark:bg-white dark:text-stone-900 dark:disabled:border-stone-700 dark:disabled:bg-stone-800 dark:disabled:text-stone-500"
             >
               Continue: {lesson.steps[index + 1].title} →
             </button>
-          ) : output?.passed && lesson.next ? (
+          ) : lesson.next ? (
             <Link
               href={lesson.next.href}
               className="rounded-lg border border-stone-900 bg-stone-900 px-4 py-2 text-sm font-semibold text-white no-underline dark:border-white dark:bg-white dark:text-stone-900"
@@ -1371,11 +533,7 @@ function AcmeStoryLab({ chapter }) {
               Next chapter: {lesson.next.title} →
             </Link>
           ) : (
-            <span className="text-sm text-stone-500">
-              {lesson.open
-                ? "Keep investigating."
-                : "Run the canonical SQL to continue."}
-            </span>
+            <span className="text-sm text-stone-500">Keep investigating.</span>
           )}
         </div>
       </section>
@@ -1413,21 +571,12 @@ export function PostgresDefaultPrivilegesLab() {
 export function PostgresOffboardingLab() {
   return <AcmeStoryLab chapter="offboarding" />;
 }
+export function PostgresMembershipMechanicsLab() {
+  return <AcmeStoryLab chapter="mechanics" />;
+}
 export function PostgresSecurityReviewLab() {
   return <AcmeStoryLab chapter="security" />;
 }
 export function PostgresAcmePlayground() {
   return <AcmeStoryLab chapter="playground" />;
 }
-
-export const acmeChapterSeeds = {
-  founderSeed,
-  directAccessSeed,
-  capabilitySeed,
-  postDriftSeed,
-  ownerSeed,
-  defaultsSeed,
-  durableSeed,
-  offboardingSeed,
-  completeSeed,
-};
