--- docs/src/components/PostgresPermissionLab.jsx
@@ -1,1712 +1,0 @@
-import { Fragment, useEffect, useRef, useState } from 'react'
-import Link from 'next/link'
-import Highlight, { defaultProps } from 'prism-react-renderer'
-
-const accessSetupStatements = [
-  'CREATE ROLE alice LOGIN',
-  'CREATE ROLE mallory LOGIN',
-  'CREATE ROLE app_owner',
-  'GRANT CREATE ON DATABASE postgres TO app_owner',
-  'SET ROLE app_owner',
-  'CREATE SCHEMA app AUTHORIZATION app_owner',
-  `CREATE TABLE app.orders (
-  id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-  customer text NOT NULL,
-  total_cents bigint NOT NULL
-)`,
-  `INSERT INTO app.orders (customer, total_cents)
-VALUES ('Acme', 4200), ('Globex', 9900)`,
-  'RESET ROLE',
-  'GRANT SELECT ON app.orders TO alice, mallory',
-  'GRANT USAGE ON SCHEMA app TO alice',
-]
-
-const hierarchySetupStatements = [
-  'CREATE ROLE alice LOGIN',
-  'CREATE ROLE bob LOGIN',
-  'CREATE ROLE team_lead',
-  'CREATE ROLE analyst',
-  'CREATE ROLE orders_reader',
-  'CREATE ROLE app_owner',
-  'GRANT CREATE ON DATABASE postgres TO app_owner',
-  'SET ROLE app_owner',
-  'CREATE SCHEMA app AUTHORIZATION app_owner',
-  `CREATE TABLE app.orders (
-  id bigint GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-  customer text NOT NULL,
-  total_cents bigint NOT NULL
-)`,
-  `INSERT INTO app.orders (customer, total_cents)
-VALUES ('Acme', 4200), ('Globex', 9900)`,
-  'RESET ROLE',
-  'GRANT USAGE ON SCHEMA app TO orders_reader',
-  'GRANT SELECT ON app.orders TO orders_reader',
-]
-
-const hierarchyExperiments = [
-  {
-    id: 'build-hierarchy',
-    title: 'Build a reusable role chain',
-    intro:
-      'If Alice inherits analyst, and analyst inherits orders_reader, do the privileges travel through both edges?',
-    explanation:
-      'Yes. PostgreSQL follows the nested membership path from Alice to analyst to orders_reader. The leaf capability owns the schema and table grants; the intermediate role groups that capability for a job.',
-    takeaway:
-      'Read GRANT orders_reader TO analyst as a directed edge: analyst may use orders_reader.',
-    actionLabel: 'Build the hierarchy',
-    path: [
-      ['Identity', 'Alice', 'pass'],
-      ['Edge 1', 'INHERIT on', 'pass'],
-      ['Bundle', 'analyst', 'focus'],
-      ['Edge 2', 'INHERIT on', 'pass'],
-      ['Capability', 'orders_reader', 'pass'],
-    ],
-    expectedResults: [{ index: 0, rowCount: 2 }],
-    highlightStatements: [0, 1],
-    statements: [
-      {
-        actor: 'postgres',
-        actorLabel: 'Lab administrator',
-        purpose: 'Grant the leaf capability',
-        sql: `GRANT orders_reader TO analyst
-  WITH INHERIT TRUE, SET TRUE`,
-      },
-      {
-        actor: 'postgres',
-        actorLabel: 'Lab administrator',
-        purpose: 'Grant the job role',
-        sql: `GRANT analyst TO alice
-  WITH INHERIT TRUE, SET TRUE`,
-      },
-      {
-        actor: 'alice',
-        actorLabel: 'Alice',
-        purpose: 'Follow both membership edges',
-        sql: 'SELECT * FROM app.orders ORDER BY id',
-      },
-    ],
-  },
-  {
-    id: 'reverse-edge',
-    title: 'Grant the role in the wrong direction',
-    intro:
-      'What happens if the same two role names are reversed in GRANT? Create the mistake instead of only reading about it.',
-    explanation:
-      'The hierarchy no longer carries orders_reader toward Alice. GRANT analyst TO orders_reader makes orders_reader a member of analyst—the opposite relationship—so Alice loses the object privileges.',
-    takeaway:
-      'GRANT capability TO member points privilege flow toward the member. Reversing the names reverses the graph.',
-    actionLabel: 'Reverse the edge',
-    path: [
-      ['Identity', 'Alice', 'pass'],
-      ['Bundle', 'analyst', 'pass'],
-      ['Edge', 'reversed', 'blocked'],
-      ['Capability', 'orders_reader', 'muted'],
-      ['Result', 'denied', 'blocked'],
-    ],
-    expectedError: 'permission denied for schema app',
-    expectedErrorCode: '42501',
-    expectedErrorIndex: 2,
-    highlightStatements: [0, 1],
-    statements: [
-      {
-        actor: 'postgres',
-        actorLabel: 'Lab administrator',
-        purpose: 'Remove the correct edge',
-        sql: 'REVOKE orders_reader FROM analyst',
-      },
-      {
-        actor: 'postgres',
-        actorLabel: 'Lab administrator',
-        purpose: 'Grant it backwards',
-        sql: 'GRANT analyst TO orders_reader',
-      },
-      {
-        actor: 'alice',
-        actorLabel: 'Alice',
-        purpose: 'Test the reversed graph',
-        sql: 'SELECT * FROM app.orders ORDER BY id',
-      },
-    ],
-  },
-  {
-    id: 'disable-inheritance',
-    title: 'Turn off automatic inheritance',
-    intro:
-      'Can Alice remain a member of analyst, retain permission to SET ROLE, but stop receiving its privileges automatically?',
-    explanation:
-      'Yes. MEMBER remains true, USAGE becomes false, and SET remains true. The edge still exists, but its ordinary privileges are dormant until Alice deliberately becomes analyst.',
-    takeaway:
-      'Membership, automatic inheritance, and permission to SET ROLE are independent facts on a PostgreSQL 16+ edge.',
-    actionLabel: 'Change the membership edge',
-    path: [
-      ['Identity', 'Alice', 'pass'],
-      ['Edge 1', 'INHERIT off', 'focus'],
-      ['Bundle', 'analyst', 'pass'],
-      ['Edge 2', 'INHERIT on', 'pass'],
-      ['Capability', 'orders_reader', 'pass'],
-    ],
-    expectedError: 'permission denied for schema app',
-    expectedErrorCode: '42501',
-    expectedErrorIndex: 5,
-    expectedResults: [
-      { index: 0, rows: [{ member: true, inherits: false, can_set: true }] },
-    ],
-    highlightStatements: [0, 1, 2, 3],
-    statements: [
-      {
-        actor: 'postgres',
-        actorLabel: 'Lab administrator',
-        purpose: 'Remove the reversed edge',
-        sql: 'REVOKE analyst FROM orders_reader',
-      },
-      {
-        actor: 'postgres',
-        actorLabel: 'Lab administrator',
-        purpose: 'Restore the capability edge',
-        sql: 'GRANT orders_reader TO analyst WITH INHERIT TRUE, SET TRUE',
-      },
-      {
-        actor: 'postgres',
-        actorLabel: 'Lab administrator',
-        purpose: 'Remove the old edge',
-        sql: 'REVOKE analyst FROM alice',
-      },
-      {
-        actor: 'postgres',
-        actorLabel: 'Lab administrator',
-        purpose: 'Recreate it without inheritance',
-        sql: `GRANT analyst TO alice
-  WITH INHERIT FALSE, SET TRUE`,
-      },
-      {
-        actor: 'alice',
-        actorLabel: 'Alice',
-        purpose: 'Inspect the edge',
-        sql: `SELECT
-  pg_has_role(current_user, 'analyst', 'MEMBER') AS member,
-  pg_has_role(current_user, 'analyst', 'USAGE')  AS inherits,
-  pg_has_role(current_user, 'analyst', 'SET')    AS can_set`,
-      },
-      {
-        actor: 'alice',
-        actorLabel: 'Alice',
-        purpose: 'Retry the read',
-        sql: 'SELECT * FROM app.orders',
-      },
-    ],
-  },
-  {
-    id: 'use-role',
-    title: 'Become the bundle deliberately',
-    intro:
-      'What changes when the effective identity becomes analyst, which still inherits orders_reader?',
-    explanation:
-      'current_user becomes analyst, so PostgreSQL evaluates the analyst hierarchy. Its inherited orders_reader privileges make the query succeed even though Alice did not inherit them automatically.',
-    takeaway:
-      'SET ROLE changes the effective identity; it does not copy grants onto the login role.',
-    actionLabel: 'Run as Alice, then SET ROLE',
-    path: [
-      ['Identity', 'analyst', 'focus'],
-      ['Membership', 'SET reported true', 'pass'],
-      ['Bundle', 'analyst', 'pass'],
-      ['Edge', 'INHERIT on', 'pass'],
-      ['Capability', 'orders_reader', 'pass'],
-    ],
-    expectedResults: [
-      {
-        index: 0,
-        rows: [{ session_user: 'alice', current_user: 'analyst' }],
-      },
-      { index: 1, rowCount: 2 },
-    ],
-    highlightStatements: [0],
-    statements: [
-      {
-        actor: 'alice',
-        actorLabel: 'Alice',
-        purpose: 'Change the effective role',
-        sql: 'SET ROLE analyst',
-      },
-      {
-        actor: 'alice',
-        actorLabel: 'Alice → analyst',
-        purpose: 'Inspect both identities',
-        sql: 'SELECT session_user, current_user',
-      },
-      {
-        actor: 'alice',
-        actorLabel: 'Alice → analyst',
-        purpose: 'Use the active hierarchy',
-        sql: 'SELECT * FROM app.orders ORDER BY id',
-      },
-    ],
-  },
-  {
-    id: 'delegate-membership',
-    title: 'Delegate membership administration',
-    intro:
-      'Can a team lead receive ADMIN without inheriting or becoming analyst, then grant analyst to Bob?',
-    explanation:
-      'Yes. ADMIN authorizes the team lead to manage membership in analyst. INHERIT and SET stay false on the team-lead edge, while Bob receives a separate inheriting edge and can read through the hierarchy.',
-    takeaway:
-      'ADMIN controls delegation, not data access. pgroles manages ADMIN and INHERIT per edge; it does not currently manage PostgreSQL’s SET option.',
-    actionLabel: 'Delegate the membership',
-    path: [
-      ['Delegator', 'team_lead', 'focus'],
-      ['Edge', 'ADMIN on', 'pass'],
-      ['New member', 'Bob', 'pass'],
-      ['Bundle', 'analyst', 'pass'],
-      ['Capability', 'orders_reader', 'pass'],
-    ],
-    expectedResults: [
-      {
-        index: 0,
-        rows: [{ admin: true, inherits: false, can_set: false }],
-      },
-      { index: 1, rowCount: 2 },
-    ],
-    highlightStatements: [1, 4],
-    statements: [
-      {
-        actor: 'postgres',
-        actorLabel: 'Lab administrator',
-        purpose: 'Create an administrative edge',
-        sql: `GRANT analyst TO team_lead
-  WITH ADMIN TRUE, INHERIT FALSE, SET FALSE`,
-      },
-      {
-        actor: 'team_lead',
-        actorLabel: 'Team lead',
-        purpose: 'Inspect the edge options',
-        sql: `SELECT
-  membership.admin_option   AS admin,
-  membership.inherit_option AS inherits,
-  membership.set_option     AS can_set
-FROM pg_auth_members AS membership
-JOIN pg_roles AS granted ON granted.oid = membership.roleid
-JOIN pg_roles AS member  ON member.oid = membership.member
-WHERE granted.rolname = 'analyst'
-  AND member.rolname = 'team_lead'`,
-      },
-      {
-        actor: 'team_lead',
-        actorLabel: 'Team lead',
-        purpose: 'Delegate analyst to Bob',
-        sql: `GRANT analyst TO bob
-  WITH INHERIT TRUE, SET TRUE`,
-      },
-      {
-        actor: 'bob',
-        actorLabel: 'Bob',
-        purpose: 'Use the delegated hierarchy',
-        sql: 'SELECT * FROM app.orders ORDER BY id',
-      },
-    ],
-  },
-]
-
-const hierarchyLesson = {
-  id: 'role-hierarchy',
-  eyebrow: 'Lesson 2 · Role hierarchy',
-  navTitle: 'Build a role hierarchy',
-  title: 'How do privileges travel through roles?',
-  description:
-    'Build a nested capability graph, control inheritance, switch effective identity, and delegate membership administration.',
-  setupStatements: hierarchySetupStatements,
-  setupSummary:
-    'six roles, a schema, the orders table, and object privileges held only by orders_reader',
-  pathLabel: 'What this step tests',
-  playgroundRoles: [
-    { value: 'alice', label: 'Alice' },
-    { value: 'bob', label: 'Bob' },
-    { value: 'analyst', label: 'Analyst' },
-    { value: 'team_lead', label: 'Team lead' },
-  ],
-  experiments: hierarchyExperiments,
-}
-const starterSql = `SELECT
-  session_user,
-  current_user,
-  current_database();`
-
-const accessQuery = 'SELECT * FROM app.orders ORDER BY id;'
-
-const pathStyles = {
-  pass: 'border-emerald-400 bg-emerald-50/70 dark:border-emerald-700 dark:bg-emerald-950/20',
-  focus:
-    'border-amber-400 bg-amber-50 dark:border-amber-500 dark:bg-amber-950/25',
-  blocked:
-    'border-rose-400 bg-rose-50 dark:border-rose-700 dark:bg-rose-950/25',
-  muted:
-    'border-stone-200 bg-stone-50 text-stone-500 dark:border-stone-700 dark:bg-stone-800/40 dark:text-stone-400',
-}
-
-function AccessPath({ path, label }) {
-  return (
-    <div className="min-w-0 max-w-full overflow-hidden rounded-2xl border border-stone-200 bg-stone-50/60 p-3 dark:border-stone-700 dark:bg-stone-950/30">
-      <p className="mb-3 mt-0 flex items-center justify-between gap-3 font-mono text-[10px] font-semibold uppercase tracking-[0.16em] text-stone-500 dark:text-stone-400">
-        <span>{label}</span>
-        <span className="sm:hidden">Swipe →</span>
-      </p>
-      <ol className="m-0 grid max-w-full list-none auto-cols-[8.5rem] grid-flow-col gap-2 overflow-x-auto p-0 pb-1 sm:auto-cols-fr sm:overflow-visible sm:pb-0">
-        {path.map(([label, value, status]) => (
-          <li
-            key={label}
-            className={`m-0 rounded-lg border-l-4 px-3 py-2 ${pathStyles[status]}`}
-          >
-            <span className="block font-mono text-[9px] font-semibold uppercase tracking-[0.12em] text-stone-500 dark:text-stone-400">
-              {label}
-            </span>
-            <span className="mt-1 block text-xs font-semibold text-stone-900 dark:text-stone-100">
-              {value}
-            </span>
-          </li>
-        ))}
-      </ol>
-    </div>
-  )
-}
-
-function SqlTokens({ code }) {
-  return (
-    <Highlight {...defaultProps} code={code} language="sql" theme={undefined}>
-      {({ tokens, getTokenProps }) =>
-        tokens.map((line, lineIndex) => (
-          <Fragment key={lineIndex}>
-            <span>
-              {line
-                .filter((token) => !token.empty)
-                .map((token, tokenIndex) => (
-                  <span key={tokenIndex} {...getTokenProps({ token })} />
-                ))}
-            </span>
-            {lineIndex < tokens.length - 1 ? '\n' : null}
-          </Fragment>
-        ))
-      }
-    </Highlight>
-  )
-}
-
-function statementSql(statement) {
-  return typeof statement === 'string' ? statement : statement.sql
-}
-
-function SqlSource({
-  statements,
-  highlightStatements = [],
-  label = 'SQL for this step',
-  maxHeight = 'max-h-[24rem]',
-}) {
-  let nextLine = 1
-
-  return (
-    <pre
-      aria-label={label}
-      className={`language-sql m-0 ${maxHeight} overflow-auto rounded-none bg-[#0c0e12] py-4 text-xs leading-5`}
-    >
-      <code>
-        {statements.map((statement, statementIndex) => {
-          const sql = statementSql(statement)
-          const code = `${sql};`
-          const startLine = nextLine
-          nextLine += code.split('\n').length
-          const highlighted = highlightStatements.includes(statementIndex)
-
-          return (
-            <Highlight
-              key={`${statementIndex}-${sql}`}
-              {...defaultProps}
-              code={code}
-              language="sql"
-              theme={undefined}
-            >
-              {({ tokens, getTokenProps }) => (
-                <span
-                  className={`block border-l-2 py-0.5 ${
-                    highlighted
-                      ? 'border-amber-400 bg-amber-300/10'
-                      : 'border-transparent'
-                  }`}
-                >
-                  {tokens.map((line, lineIndex) => (
-                    <Fragment key={lineIndex}>
-                      <span className="grid grid-cols-[2.75rem_1fr] px-3 sm:px-4">
-                        <span
-                          aria-hidden="true"
-                          className={`select-none pr-3 text-right font-mono ${
-                            highlighted ? 'text-amber-400' : 'text-stone-600'
-                          }`}
-                        >
-                          {startLine + lineIndex}
-                        </span>
-                        <span className="min-w-0 whitespace-pre-wrap break-words">
-                          {line
-                            .filter((token) => !token.empty)
-                            .map((token, tokenIndex) => (
-                              <span
-                                key={tokenIndex}
-                                {...getTokenProps({ token })}
-                              />
-                            ))}
-                        </span>
-                      </span>
-                      {lineIndex === tokens.length - 1 && (
-                        <span className="block h-3" />
-                      )}
-                    </Fragment>
-                  ))}
-                </span>
-              )}
-            </Highlight>
-          )
-        })}
-      </code>
-    </pre>
-  )
-}
-
-function SqlEditor({
-  value,
-  onChange,
-  disabled,
-  ariaLabel = 'SQL playground',
-  minHeight = 'min-h-[11rem]',
-}) {
-  const highlightRef = useRef(null)
-
-  function syncScroll(event) {
-    if (!highlightRef.current) return
-    highlightRef.current.scrollTop = event.currentTarget.scrollTop
-    highlightRef.current.scrollLeft = event.currentTarget.scrollLeft
-  }
-
-  return (
-    <div
-      className={`relative ${minHeight} overflow-hidden rounded-xl border border-stone-700 bg-[#0c0e12] focus-within:border-amber-400 focus-within:ring-2 focus-within:ring-amber-400/20`}
-    >
-      <pre
-        ref={highlightRef}
-        aria-hidden="true"
-        className={`pointer-events-none absolute inset-0 m-0 ${minHeight} overflow-hidden whitespace-pre p-4 font-mono text-xs leading-6 text-stone-200`}
-      >
-        <code>
-          <SqlTokens code={value || ' '} />
-        </code>
-      </pre>
-      <textarea
-        aria-label={ariaLabel}
-        value={value}
-        onChange={(event) => onChange(event.target.value)}
-        onScroll={syncScroll}
-        disabled={disabled}
-        wrap="off"
-        spellCheck="false"
-        className={`relative z-10 m-0 ${minHeight} w-full resize-y overflow-auto border-0 bg-transparent p-4 font-mono text-xs leading-6 text-transparent outline-none selection:bg-amber-300/30 disabled:cursor-wait`}
-        style={{ caretColor: '#f5f5f4', WebkitTextFillColor: 'transparent' }}
-      />
-    </div>
-  )
-}
-
-function valueText(value) {
-  if (value === null) return 'NULL'
-  if (typeof value === 'boolean') return value ? 'true' : 'false'
-  return String(value)
-}
-
-function resultsMatchExpectations(results, expectations = []) {
-  return expectations.every((expectation) => {
-    const result = results[expectation.index]
-    if (!result) return false
-    if (
-      expectation.rowCount !== undefined &&
-      result.rows.length !== expectation.rowCount
-    )
-      return false
-    if (!expectation.rows) return true
-    if (result.rows.length !== expectation.rows.length) return false
-
-    return expectation.rows.every((expectedRow, rowIndex) =>
-      Object.entries(expectedRow).every(
-        ([field, value]) => result.rows[rowIndex]?.[field] === value
-      )
-    )
-  })
-}
-
-function ResultTable({ entry }) {
-  return (
-    <div className="overflow-x-auto rounded-xl border border-stone-700">
-      <table className="min-w-full border-collapse font-mono text-xs">
-        <thead className="bg-stone-800 text-left text-stone-300">
-          <tr>
-            {entry.fields.map((field) => (
-              <th
-                key={field}
-                className="border-b border-stone-700 px-3 py-2 font-medium"
-              >
-                {field}
-              </th>
-            ))}
-          </tr>
-        </thead>
-        <tbody>
-          {entry.rows.length === 0 ? (
-            <tr>
-              <td
-                colSpan={entry.fields.length}
-                className="px-3 py-3 text-stone-500"
-              >
-                (0 rows)
-              </td>
-            </tr>
-          ) : (
-            entry.rows.map((row, rowIndex) => (
-              <tr
-                key={rowIndex}
-                className="border-b border-stone-800 last:border-0"
-              >
-                {entry.fields.map((field) => (
-                  <td key={field} className="px-3 py-2 text-stone-200">
-                    {valueText(row[field])}
-                  </td>
-                ))}
-              </tr>
-            ))
-          )}
-        </tbody>
-      </table>
-    </div>
-  )
-}
-
-function SqlOutput({
-  output,
-  operation,
-  kind = 'lesson',
-  expectedError = false,
-}) {
-  const runningThisOutput =
-    operation === kind || operation === 'booting' || operation === 'preparing'
-
-  if (operation === 'booting') {
-    return (
-      <p className="m-0 text-sm text-stone-300">
-        Downloading and starting PostgreSQL 18.3…
-      </p>
-    )
-  }
-  if (operation === 'preparing') {
-    return (
-      <p className="m-0 text-sm text-stone-300">
-        Preparing this lesson’s roles and objects…
-      </p>
-    )
-  }
-  if (runningThisOutput) {
-    return <p className="m-0 text-sm text-stone-300">Executing SQL…</p>
-  }
-  if (!output) {
-    return (
-      <p className="m-0 text-sm leading-6 text-stone-400">
-        PostgreSQL’s command tags, rows, and exact error will appear here.
-      </p>
-    )
-  }
-
-  return (
-    <div className="space-y-4">
-      {output.commands.length > 0 && (
-        <div
-          className="flex flex-wrap gap-1.5"
-          aria-label="PostgreSQL command results"
-        >
-          {output.commands.map((command, index) => (
-            <span
-              key={`${command}-${index}`}
-              className="rounded-md border border-stone-700 bg-stone-800 px-2 py-1 font-mono text-[10px] text-stone-300"
-            >
-              {command}
-            </span>
-          ))}
-        </div>
-      )}
-      {output.results.map((entry, index) => (
-        <ResultTable key={index} entry={entry} />
-      ))}
-      {output.error && (
-        <div
-          className={`rounded-xl border px-4 py-3 font-mono text-xs ${
-            expectedError
-              ? 'border-amber-700/70 bg-amber-950/35 text-amber-200'
-              : 'border-red-800 bg-red-950/40 text-red-200'
-          }`}
-        >
-          <span className="font-semibold">ERROR:</span> {output.error}
-        </div>
-      )}
-      {kind !== 'lesson' && !output.error ? (
-        <p className="m-0 text-sm font-semibold text-emerald-300">
-          ✓ SQL completed
-        </p>
-      ) : null}
-    </div>
-  )
-}
-
-function GuidedSqlNotebook({ experiment, output, operation }) {
-  const running = operation === 'lesson'
-
-  return (
-    <div className="overflow-hidden rounded-2xl border border-stone-700 bg-[#0c0e12]">
-      {experiment.statements.map((statement, index) => {
-        const entry = output?.entries?.[index]
-        const highlighted = experiment.highlightStatements.includes(index)
-
-        return (
-          <section
-            key={`${experiment.id}-${index}`}
-            className={`border-b border-stone-700 last:border-b-0 ${
-              highlighted ? 'bg-amber-300/[0.04]' : ''
-            }`}
-            aria-label={`${statement.purpose}, run as ${statement.actorLabel}`}
-          >
-            <div className="flex flex-wrap items-center justify-between gap-2 border-b border-stone-800 bg-stone-900/80 px-4 py-2.5">
-              <span className="font-mono text-[10px] font-semibold uppercase tracking-[0.14em] text-stone-400">
-                {statement.purpose}
-              </span>
-              <span
-                className={`rounded-full border px-2.5 py-1 font-mono text-[10px] font-semibold ${
-                  highlighted
-                    ? 'border-amber-500/60 bg-amber-400/10 text-amber-200'
-                    : 'border-stone-700 bg-stone-800 text-stone-300'
-                }`}
-              >
-                Run as {statement.actorLabel}
-              </span>
-            </div>
-            <SqlSource
-              statements={[statement]}
-              highlightStatements={highlighted ? [0] : []}
-              label={`${statement.purpose} SQL`}
-              maxHeight="max-h-[18rem]"
-            />
-            {(entry || running) && (
-              <div className="border-t border-stone-800 bg-stone-950 px-4 py-4 text-stone-200">
-                <p className="mb-3 mt-0 font-mono text-[9px] font-semibold uppercase tracking-[0.16em] text-stone-500">
-                  {entry ? 'Result' : 'Running'}
-                </p>
-                {entry ? (
-                  <SqlOutput
-                    output={entry}
-                    operation="idle"
-                    expectedError={
-                      output?.passed && experiment.expectedErrorIndex === index
-                    }
-                  />
-                ) : (
-                  <p className="m-0 text-sm text-stone-300">Executing SQL…</p>
-                )}
-              </div>
-            )}
-          </section>
-        )
-      })}
-      {output && !output.passed && (
-        <div className="border-t border-red-900/70 bg-red-950/35 px-4 py-3 text-sm text-red-200">
-          This step did not reach its starting-state expectation. Restore the
-          lesson and try again.
-        </div>
-      )}
-    </div>
-  )
-}
-
-function appendStatementResults(target, statementResults) {
-  for (const result of statementResults) {
-    if (result.command) target.commands.push(result.command)
-    if (result.fields?.length > 0) {
-      target.results.push({
-        fields: result.fields.map((field) => field.name),
-        rows: result.rows,
-      })
-    }
-  }
-}
-
-async function executeStatements(database, statements) {
-  const output = { commands: [], results: [], entries: [], error: null }
-  let activeActor = null
-
-  for (let index = 0; index < statements.length; index += 1) {
-    const statement = statements[index]
-    const sql = statementSql(statement)
-    const actor = typeof statement === 'string' ? null : statement.actor
-    const entry = {
-      sql,
-      actor,
-      commands: [],
-      results: [],
-      error: null,
-    }
-    output.entries.push(entry)
-
-    try {
-      if (actor && actor !== activeActor) {
-        await database.exec('RESET ROLE;')
-        await database.exec(
-          `SET SESSION AUTHORIZATION "${actor.replaceAll('"', '""')}";`
-        )
-        activeActor = actor
-      }
-
-      const statementResults = await database.exec(`${sql};`)
-      appendStatementResults(output, statementResults)
-      appendStatementResults(entry, statementResults)
-    } catch (error) {
-      output.error = error.message
-      output.errorCode = error.code
-      output.errorIndex = index
-      entry.error = error.message
-      entry.errorCode = error.code
-      break
-    }
-  }
-
-  try {
-    await database.exec('RESET ROLE;')
-    await database.exec('SET SESSION AUTHORIZATION postgres;')
-  } catch {
-    // Guided SQL never receives superuser access, so this only guards teardown.
-  }
-
-  return output
-}
-
-async function executeSetupStatements(database, statements) {
-  const output = { commands: [], results: [], error: null }
-
-  for (let index = 0; index < statements.length; index += 1) {
-    try {
-      const statementResults = await database.exec(`${statements[index]};`)
-      appendStatementResults(output, statementResults)
-    } catch (error) {
-      output.error = error.message
-      output.errorCode = error.code
-      output.errorIndex = index
-      break
-    }
-  }
-
-  return output
-}
-
-function normalizeSql(sql) {
-  return sql
-    .trim()
-    .replace(/;+\s*$/, '')
-    .replace(/\s+/g, ' ')
-    .toLowerCase()
-}
-
-function AccessStoryLab() {
-  const databaseRef = useRef(null)
-  const databasePromiseRef = useRef(null)
-  const lessonPreparedRef = useRef(false)
-  const operationLockRef = useRef(false)
-  const [actor, setActor] = useState('alice')
-  const [sql, setSql] = useState(accessQuery)
-  const [output, setOutput] = useState(null)
-  const [operation, setOperation] = useState('idle')
-  const [observed, setObserved] = useState({ alice: false, mallory: false })
-  const running = operation !== 'idle'
-  const exampleRestored = normalizeSql(sql) === normalizeSql(accessQuery)
-
-  useEffect(
-    () => () => {
-      databaseRef.current?.close().catch(() => {})
-    },
-    []
-  )
-
-  async function ensureDatabase() {
-    if (databaseRef.current) return databaseRef.current
-    if (databasePromiseRef.current) return databasePromiseRef.current
-
-    setOperation('booting')
-    databasePromiseRef.current = import('@electric-sql/pglite')
-      .then(({ PGlite }) => PGlite.create())
-      .then((database) => {
-        databaseRef.current = database
-        return database
-      })
-      .finally(() => {
-        databasePromiseRef.current = null
-      })
-
-    return databasePromiseRef.current
-  }
-
-  async function ensureLessonPrepared() {
-    const database = await ensureDatabase()
-    if (lessonPreparedRef.current) return database
-
-    setOperation('preparing')
-    const setupOutput = await executeSetupStatements(
-      database,
-      accessSetupStatements
-    )
-    if (setupOutput.error) {
-      throw new Error(`Lesson setup failed: ${setupOutput.error}`)
-    }
-    lessonPreparedRef.current = true
-    return database
-  }
-
-  async function runQuery() {
-    if (operationLockRef.current || running) return
-
-    const selectedActor = actor
-    const selectedSql = sql
-    if (selectedSql.trim() === '') return
-
-    setOutput(null)
-    operationLockRef.current = true
-    let sandbox = null
-
-    try {
-      const database = await ensureLessonPrepared()
-      setOperation('query')
-      sandbox = await database.clone()
-      await sandbox.exec(
-        `SET SESSION AUTHORIZATION "${selectedActor.replaceAll('"', '""')}";`
-      )
-
-      const result = { commands: [], results: [], error: null }
-      try {
-        const statementResults = await sandbox.exec(selectedSql)
-        appendStatementResults(result, statementResults)
-      } catch (error) {
-        result.error = error.message
-        result.errorCode = error.code
-      }
-
-      const canonical = normalizeSql(selectedSql) === normalizeSql(accessQuery)
-      const returnedOrders =
-        selectedActor === 'alice' &&
-        canonical &&
-        !result.error &&
-        result.results[0]?.rows?.length === 2
-      const exposedSchemaGate =
-        selectedActor === 'mallory' &&
-        canonical &&
-        result.errorCode === '42501' &&
-        result.error?.includes('permission denied for schema app')
-      const guidedOutcome = returnedOrders
-        ? 'alice-success'
-        : exposedSchemaGate
-        ? 'mallory-denied'
-        : null
-
-      setOutput({
-        ...result,
-        actor: selectedActor,
-        sql: selectedSql,
-        canonical,
-        guidedOutcome,
-      })
-
-      if (returnedOrders) {
-        setObserved((current) => ({ ...current, alice: true }))
-      }
-      if (exposedSchemaGate) {
-        setObserved((current) => ({ ...current, mallory: true }))
-      }
-    } catch (error) {
-      setOutput({
-        commands: [],
-        results: [],
-        error: error.message,
-        actor: selectedActor,
-        sql: selectedSql,
-        canonical: false,
-        guidedOutcome: null,
-      })
-    } finally {
-      if (sandbox) await sandbox.close().catch(() => {})
-      setOperation('idle')
-      operationLockRef.current = false
-    }
-  }
-
-  function restoreExample() {
-    if (running) return
-    setSql(accessQuery)
-    setOutput(null)
-  }
-
-  const decisionPath =
-    output?.guidedOutcome === 'alice-success'
-      ? [
-          ['Schema', 'USAGE granted', 'pass'],
-          ['Table', 'SELECT granted', 'pass'],
-          ['Decision', '2 rows returned', 'pass'],
-        ]
-      : output?.guidedOutcome === 'mallory-denied'
-      ? [
-          ['Schema', 'USAGE missing', 'blocked'],
-          ['Table · not reached', 'SELECT present', 'muted'],
-          ['Decision', 'Access denied', 'blocked'],
-        ]
-      : null
-
-  return (
-    <div className="not-prose my-10 overflow-hidden rounded-[2rem] border border-stone-300/90 bg-white shadow-[0_30px_80px_-52px_rgba(28,25,23,0.5)] dark:border-stone-700 dark:bg-stone-900 dark:shadow-none">
-      <div className="hidden border-b border-stone-300/80 bg-[linear-gradient(135deg,rgba(254,243,199,0.75),rgba(240,253,250,0.8))] px-5 py-6 dark:border-stone-700 dark:bg-[linear-gradient(135deg,rgba(120,53,15,0.2),rgba(19,78,74,0.2))] sm:block sm:px-7">
-        <div className="flex flex-wrap items-center justify-between gap-3">
-          <p className="m-0 text-[11px] font-semibold uppercase tracking-[0.22em] text-amber-800 dark:text-amber-300">
-            Live PostgreSQL · The orders report
-          </p>
-          <a
-            href="https://pglite.dev/"
-            className="rounded-full border border-stone-300 bg-white/75 px-3 py-1 font-mono text-[10px] font-semibold text-stone-700 no-underline shadow-none dark:border-stone-700 dark:bg-stone-900/70 dark:text-stone-300"
-          >
-            PostgreSQL 18.3 · PGlite ↗
-          </a>
-        </div>
-        <h2 className="mt-3 font-display text-2xl tracking-[-0.02em] text-stone-950 dark:text-white">
-          One query, two permission paths
-        </h2>
-        <p className="mb-0 mt-2 max-w-3xl text-sm leading-6 text-stone-700 dark:text-stone-300">
-          Alice is an approved reader. Mallory is not—but the database has
-          picked up one direct grant outside policy.
-        </p>
-      </div>
-
-      <div className="flex items-center justify-between gap-3 border-b border-stone-300/80 bg-[linear-gradient(135deg,rgba(254,243,199,0.75),rgba(240,253,250,0.8))] px-5 py-4 dark:border-stone-700 dark:bg-[linear-gradient(135deg,rgba(120,53,15,0.2),rgba(19,78,74,0.2))] sm:hidden">
-        <h2 className="m-0 text-[10px] font-semibold uppercase tracking-[0.18em] text-amber-800 dark:text-amber-300">
-          Live PostgreSQL · orders report
-        </h2>
-        <span className="shrink-0 rounded-full border border-stone-300 bg-white/75 px-2.5 py-1 font-mono text-[9px] font-semibold text-stone-700 dark:border-stone-700 dark:bg-stone-900/70 dark:text-stone-300">
-          PG 18.3
-        </span>
-      </div>
-
-      <section className="grid gap-5 p-5 sm:p-7" aria-labelledby="query-lab">
-        <div>
-          <p className="m-0 font-mono text-[10px] font-semibold uppercase tracking-[0.16em] text-amber-700 dark:text-amber-300">
-            Establish the baseline
-          </p>
-          <h3
-            id="query-lab"
-            className="mt-2 font-display text-2xl tracking-[-0.02em] text-stone-950 dark:text-white"
-          >
-            Start with the report working
-          </h3>
-          <p className="mb-0 mt-3 max-w-3xl text-base leading-7 text-stone-700 dark:text-stone-300">
-            Alice has the complete access path. Send the application query and
-            watch PostgreSQL return the orders.
-          </p>
-        </div>
-
-        <div className="overflow-hidden rounded-2xl border border-stone-700 bg-[#0c0e12]">
-          <div className="flex items-end justify-between gap-2 border-b border-stone-700 bg-stone-900 px-4 py-3">
-            <label className="grid min-w-0 flex-1 gap-1.5 font-mono text-[10px] font-semibold uppercase tracking-[0.14em] text-stone-400 sm:flex-none">
-              Query as
-              <select
-                value={actor}
-                onChange={(event) => {
-                  setActor(event.target.value)
-                  setOutput(null)
-                }}
-                disabled={running}
-                className="w-full min-w-0 rounded-lg border border-stone-600 bg-stone-950 px-3 py-2 font-mono text-sm normal-case tracking-normal text-white sm:min-w-40"
-              >
-                <option value="alice">Alice</option>
-                <option value="mallory">Mallory</option>
-              </select>
-            </label>
-            <div className="flex flex-wrap gap-2">
-              {!exampleRestored && (
-                <button
-                  type="button"
-                  onClick={restoreExample}
-                  disabled={running}
-                  className="rounded-lg border border-stone-600 px-3 py-2 text-xs font-semibold text-stone-300 hover:border-amber-400 hover:text-white disabled:cursor-wait"
-                >
-                  Restore query
-                </button>
-              )}
-              <button
-                type="button"
-                onClick={() => runQuery()}
-                disabled={running || sql.trim() === ''}
-                className="rounded-xl border border-amber-300 bg-amber-300 px-5 py-2.5 font-mono text-xs font-bold text-stone-950 transition hover:bg-amber-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-amber-300 focus-visible:ring-offset-2 focus-visible:ring-offset-stone-900 disabled:cursor-wait disabled:border-stone-600 disabled:bg-stone-700 disabled:text-stone-400"
-              >
-                {operation === 'booting'
-                  ? 'Starting PostgreSQL…'
-                  : operation === 'preparing'
-                  ? 'Preparing database…'
-                  : operation === 'query'
-                  ? 'Running query…'
-                  : 'Run query'}
-              </button>
-            </div>
-          </div>
-
-          <div className="p-3 sm:p-4">
-            <SqlEditor
-              value={sql}
-              onChange={(value) => {
-                setSql(value)
-                setOutput(null)
-              }}
-              disabled={running}
-              ariaLabel="Orders report SQL"
-              minHeight="min-h-[6rem]"
-            />
-          </div>
-
-          <div className="border-t border-stone-800 bg-stone-950 px-4 py-3">
-            <p className="m-0 text-xs leading-5 text-stone-400">
-              Temporary database copy · changes are discarded after each run
-            </p>
-          </div>
-
-          <div
-            className="border-t border-stone-800 bg-stone-950 px-4 py-5 text-stone-200"
-            aria-live="polite"
-            aria-busy={running}
-          >
-            <p className="mb-4 mt-0 font-mono text-[10px] font-semibold uppercase tracking-[0.16em] text-stone-500">
-              PostgreSQL output
-            </p>
-            <SqlOutput
-              output={output}
-              operation={operation}
-              kind="query"
-              expectedError={output?.guidedOutcome === 'mallory-denied'}
-            />
-          </div>
-        </div>
-
-        {decisionPath && (
-          <AccessPath path={decisionPath} label="Why PostgreSQL decided this" />
-        )}
-
-        {output && !output.guidedOutcome && (
-          <p className="m-0 rounded-xl border border-stone-200 bg-stone-50 px-4 py-3 text-sm leading-6 text-stone-600 dark:border-stone-700 dark:bg-stone-950/30 dark:text-stone-300">
-            That is PostgreSQL’s real response to your SQL. Restore the example
-            query to continue the guided comparison.
-          </p>
-        )}
-
-        {!observed.alice && (
-          <p className="m-0 border-l-2 border-amber-400 pl-4 text-sm leading-6 text-stone-700 dark:text-stone-300">
-            Begin with Alice so the successful access path has a baseline.
-          </p>
-        )}
-
-        {observed.alice && !observed.mallory && (
-          <div className="rounded-2xl border border-amber-200 bg-amber-50/70 p-5 dark:border-amber-900/70 dark:bg-amber-950/25">
-            <p className="m-0 font-mono text-[10px] font-semibold uppercase tracking-[0.16em] text-amber-800 dark:text-amber-300">
-              Now change one variable
-            </p>
-            <h3 className="mb-0 mt-2 font-display text-xl text-stone-950 dark:text-white">
-              Mallory has a stray table grant
-            </h3>
-            <p className="mb-0 mt-3 text-sm leading-6 text-stone-700 dark:text-stone-300">
-              Someone granted Mallory <code>SELECT</code> directly on{' '}
-              <code>app.orders</code>, even though she is not an approved
-              reader. Keep the database and query unchanged; change only who is
-              asking.
-            </p>
-            <p className="mb-0 mt-3 text-sm font-semibold leading-6 text-stone-900 dark:text-white">
-              Choose Mallory under Query as, then run the unchanged query again.
-            </p>
-          </div>
-        )}
-
-        {observed.alice && observed.mallory && (
-          <div className="rounded-2xl border border-emerald-200 bg-emerald-50/70 p-5 dark:border-emerald-900 dark:bg-emerald-950/25">
-            <p className="m-0 font-mono text-[10px] font-semibold uppercase tracking-[0.16em] text-emerald-700 dark:text-emerald-300">
-              What the comparison revealed
-            </p>
-            <p className="mb-0 mt-2 text-sm leading-6 text-stone-800 dark:text-stone-200">
-              Mallory’s table grant is real, but PostgreSQL stops first at the
-              missing schema permission. That partial grant is still dangerous
-              drift: another membership or <code>PUBLIC</code> grant could
-              complete the path later.
-            </p>
-            <p className="mb-0 mt-3 border-l-2 border-emerald-500 pl-3 text-sm font-semibold leading-6 text-stone-900 dark:text-white">
-              The production fix is to remove Mallory’s undeclared grant and
-              declare the complete reader path once.
-            </p>
-            <Link
-              href="#make-the-access-path-durable"
-              className="mt-4 inline-flex rounded-lg border border-emerald-700 px-4 py-2 text-sm font-semibold text-emerald-800 no-underline hover:bg-emerald-100 dark:border-emerald-600 dark:text-emerald-200 dark:hover:bg-emerald-950"
-            >
-              See the pgroles policy ↓
-            </Link>
-          </div>
-        )}
-      </section>
-
-      <details className="border-t border-stone-200 px-5 py-4 dark:border-stone-800 sm:px-7">
-        <summary className="cursor-pointer text-sm font-semibold text-stone-600 hover:text-stone-950 dark:text-stone-300 dark:hover:text-white">
-          How role selection works in the browser
-        </summary>
-        <p className="mb-0 mt-3 text-sm leading-6 text-stone-500 dark:text-stone-400">
-          The selector sets the starting session authorization inside an
-          isolated browser database. Each query runs in a disposable copy, so
-          arbitrary SQL cannot alter the guided scenario. It is not an
-          authentication test and does not exercise passwords,{' '}
-          <code>CONNECT</code>, <code>LOGIN</code>, or <code>pg_hba.conf</code>.
-        </p>
-      </details>
-
-      <details className="border-t border-stone-200 px-5 py-4 dark:border-stone-800 sm:px-7">
-        <summary className="cursor-pointer text-sm font-semibold text-stone-600 hover:text-stone-950 dark:text-stone-300 dark:hover:text-white">
-          Show the database setup SQL
-        </summary>
-        <p className="mb-4 mt-3 text-sm leading-6 text-stone-500 dark:text-stone-400">
-          The setup deliberately gives Mallory a table grant without schema
-          access so the comparison changes only the session identity.
-        </p>
-        <div className="overflow-hidden rounded-2xl border border-stone-700">
-          <SqlSource
-            statements={accessSetupStatements}
-            label="Lesson setup SQL"
-            maxHeight="max-h-[24rem]"
-          />
-        </div>
-      </details>
-    </div>
-  )
-}
-
-function PostgresLesson({ lesson, nextLesson }) {
-  const databaseRef = useRef(null)
-  const databasePromiseRef = useRef(null)
-  const lessonPreparedRef = useRef(false)
-  const operationLockRef = useRef(false)
-  const lessonStepRef = useRef(null)
-  const [activeIndex, setActiveIndex] = useState(0)
-  const [completed, setCompleted] = useState([])
-  const [outputs, setOutputs] = useState({})
-  const [operation, setOperation] = useState('idle')
-  const [playgroundOpen, setPlaygroundOpen] = useState(false)
-  const [playgroundSql, setPlaygroundSql] = useState(starterSql)
-  const [playgroundRole, setPlaygroundRole] = useState(
-    lesson.playgroundRoles[0].value
-  )
-  const [playgroundOutput, setPlaygroundOutput] = useState(null)
-  const experiments = lesson.experiments
-  const experiment = experiments[activeIndex]
-  const output = outputs[experiment.id]
-  const running = operation !== 'idle'
-
-  useEffect(
-    () => () => {
-      databaseRef.current?.close().catch(() => {})
-    },
-    []
-  )
-
-  async function ensureDatabase() {
-    if (databaseRef.current) return databaseRef.current
-    if (databasePromiseRef.current) return databasePromiseRef.current
-
-    setOperation('booting')
-    databasePromiseRef.current = import('@electric-sql/pglite')
-      .then(({ PGlite }) => PGlite.create())
-      .then((database) => {
-        databaseRef.current = database
-        return database
-      })
-      .finally(() => {
-        databasePromiseRef.current = null
-      })
-
-    return databasePromiseRef.current
-  }
-
-  async function ensureLessonPrepared() {
-    const database = await ensureDatabase()
-    if (lessonPreparedRef.current) return database
-
-    setOperation('preparing')
-    const setupOutput = await executeSetupStatements(
-      database,
-      lesson.setupStatements
-    )
-    if (setupOutput.error) {
-      throw new Error(`Lesson setup failed: ${setupOutput.error}`)
-    }
-    lessonPreparedRef.current = true
-    return database
-  }
-
-  async function runExperiment() {
-    if (operationLockRef.current || running || output) return
-    operationLockRef.current = true
-
-    try {
-      const database = await ensureLessonPrepared()
-      setOperation('lesson')
-      const result = await executeStatements(database, experiment.statements)
-      const expectedErrorMatched = Boolean(
-        experiment.expectedError &&
-          result.errorIndex === experiment.expectedErrorIndex &&
-          result.error?.includes(experiment.expectedError) &&
-          (!experiment.expectedErrorCode ||
-            result.errorCode === experiment.expectedErrorCode)
-      )
-      const resultValuesMatch = resultsMatchExpectations(
-        result.results,
-        experiment.expectedResults
-      )
-      const passed =
-        (experiment.expectedError
-          ? expectedErrorMatched
-          : result.error === null) && resultValuesMatch
-      const lessonOutput = { ...result, passed }
-
-      setOutputs((current) => ({ ...current, [experiment.id]: lessonOutput }))
-      if (passed) {
-        setCompleted((current) =>
-          current.includes(experiment.id)
-            ? current
-            : [...current, experiment.id]
-        )
-      }
-    } catch (error) {
-      setOutputs((current) => ({
-        ...current,
-        [experiment.id]: {
-          commands: [],
-          results: [],
-          entries: [],
-          error: error.message,
-          passed: false,
-        },
-      }))
-    } finally {
-      setOperation('idle')
-      operationLockRef.current = false
-    }
-  }
-
-  async function runPlaygroundSql() {
-    if (operationLockRef.current || running || playgroundSql.trim() === '')
-      return
-
-    operationLockRef.current = true
-    setPlaygroundOutput(null)
-    let sandbox = null
-
-    try {
-      const database = await ensureLessonPrepared()
-      setOperation('playground')
-      sandbox = await database.clone()
-      await sandbox.exec(
-        `SET SESSION AUTHORIZATION "${playgroundRole.replaceAll('"', '""')}";`
-      )
-      const result = { commands: [], results: [], error: null }
-      try {
-        const statementResults = await sandbox.exec(playgroundSql)
-        appendStatementResults(result, statementResults)
-      } catch (error) {
-        result.error = error.message
-        result.errorCode = error.code
-      }
-      setPlaygroundOutput(result)
-    } catch (error) {
-      setPlaygroundOutput({ commands: [], results: [], error: error.message })
-    } finally {
-      if (sandbox) await sandbox.close().catch(() => {})
-      setOperation('idle')
-      operationLockRef.current = false
-    }
-  }
-
-  async function resetLab() {
-    if (operationLockRef.current || running) return
-    operationLockRef.current = true
-    setOperation('resetting')
-    const database = databaseRef.current
-    databaseRef.current = null
-    databasePromiseRef.current = null
-    lessonPreparedRef.current = false
-
-    try {
-      if (database) await database.close()
-    } catch {
-      // The old instance is detached; the next action can still create a fresh database.
-    } finally {
-      setCompleted([])
-      setOutputs({})
-      setActiveIndex(0)
-      setPlaygroundOpen(false)
-      setPlaygroundOutput(null)
-      setPlaygroundSql(starterSql)
-      setPlaygroundRole(lesson.playgroundRoles[0].value)
-      setOperation('idle')
-      operationLockRef.current = false
-    }
-  }
-
-  function showPrevious() {
-    if (running || activeIndex === 0) return
-    setPlaygroundOutput(null)
-    setActiveIndex((index) => index - 1)
-    requestAnimationFrame(() =>
-      lessonStepRef.current?.scrollIntoView({
-        behavior: 'smooth',
-        block: 'start',
-      })
-    )
-  }
-
-  function showNext() {
-    if (running || !output?.passed || activeIndex === experiments.length - 1)
-      return
-    setPlaygroundOutput(null)
-    setActiveIndex((index) => index + 1)
-    requestAnimationFrame(() =>
-      lessonStepRef.current?.scrollIntoView({
-        behavior: 'smooth',
-        block: 'start',
-      })
-    )
-  }
-
-  return (
-    <div className="not-prose my-10 overflow-hidden rounded-[2rem] border border-stone-300/90 bg-white shadow-[0_30px_80px_-52px_rgba(28,25,23,0.5)] dark:border-stone-700 dark:bg-stone-900 dark:shadow-none">
-      <div className="hidden border-b border-stone-300/80 bg-[linear-gradient(135deg,rgba(254,243,199,0.75),rgba(240,253,250,0.8))] px-5 py-6 dark:border-stone-700 dark:bg-[linear-gradient(135deg,rgba(120,53,15,0.2),rgba(19,78,74,0.2))] sm:block sm:px-7">
-        <div className="flex flex-wrap items-center justify-between gap-3">
-          <p className="m-0 text-[11px] font-semibold uppercase tracking-[0.22em] text-amber-800 dark:text-amber-300">
-            {lesson.eyebrow}
-          </p>
-          <a
-            href="https://pglite.dev/"
-            className="rounded-full border border-stone-300 bg-white/75 px-3 py-1 font-mono text-[10px] font-semibold text-stone-700 no-underline shadow-none dark:border-stone-700 dark:bg-stone-900/70 dark:text-stone-300"
-          >
-            PostgreSQL 18.3 · PGlite ↗
-          </a>
-        </div>
-        <h2 className="mt-3 font-display text-2xl tracking-[-0.02em] text-stone-950 dark:text-white">
-          {lesson.title}
-        </h2>
-        <p className="mb-0 mt-2 max-w-3xl text-sm leading-6 text-stone-700 dark:text-stone-300">
-          {lesson.description}
-        </p>
-      </div>
-
-      <div className="border-b border-stone-200 px-5 py-5 dark:border-stone-800 sm:px-7">
-        <div className="flex items-start justify-between gap-5">
-          <div>
-            <p className="m-0 font-mono text-[10px] font-semibold uppercase tracking-[0.16em] text-amber-700 dark:text-amber-300">
-              {lesson.navTitle}
-            </p>
-            <p className="mb-0 mt-1 text-sm text-stone-600 dark:text-stone-300">
-              Step {activeIndex + 1} of {experiments.length}
-            </p>
-          </div>
-          <button
-            type="button"
-            onClick={resetLab}
-            disabled={running}
-            className="shrink-0 rounded-lg border border-stone-300 px-3 py-1.5 font-mono text-xs font-semibold text-stone-600 hover:border-amber-400 hover:text-stone-950 disabled:cursor-wait disabled:opacity-50 dark:border-stone-700 dark:text-stone-300 dark:hover:text-white"
-          >
-            Reset lesson
-          </button>
-        </div>
-        <ol
-          className="mt-4 flex gap-1.5"
-          aria-label={`Step ${activeIndex + 1} of ${experiments.length}`}
-        >
-          {experiments.map((item, index) => (
-            <li key={item.id} className="flex-1">
-              <span
-                className={`block h-1.5 rounded-full ${
-                  completed.includes(item.id)
-                    ? 'bg-emerald-500'
-                    : index === activeIndex
-                    ? 'bg-amber-400'
-                    : 'bg-stone-200 dark:bg-stone-700'
-                }`}
-              >
-                <span className="sr-only">
-                  {item.title}
-                  {completed.includes(item.id)
-                    ? ': complete'
-                    : index === activeIndex
-                    ? ': current'
-                    : ''}
-                </span>
-              </span>
-            </li>
-          ))}
-        </ol>
-      </div>
-
-      <section
-        ref={lessonStepRef}
-        className="grid scroll-mt-24 gap-5 p-5 sm:p-7"
-        aria-labelledby={`experiment-${experiment.id}`}
-      >
-        <div>
-          <p className="m-0 font-mono text-[10px] font-semibold uppercase tracking-[0.16em] text-stone-500 dark:text-stone-400">
-            Step {activeIndex + 1}
-          </p>
-          <h3
-            id={`experiment-${experiment.id}`}
-            className="mt-2 font-display text-2xl tracking-[-0.02em] text-stone-950 dark:text-white"
-          >
-            {experiment.title}
-          </h3>
-          <p className="mb-0 mt-3 max-w-3xl text-base leading-7 text-stone-700 dark:text-stone-300">
-            {experiment.intro}
-          </p>
-        </div>
-
-        <div className="grid gap-3">
-          <div className="flex flex-col gap-4 rounded-2xl border border-stone-700 bg-stone-900 px-4 py-3 sm:flex-row sm:items-center sm:justify-between">
-            <p className="m-0 font-mono text-[10px] font-semibold uppercase tracking-[0.14em] text-stone-400">
-              <span className="text-amber-300">Amber</span> marks the focus of
-              this step
-            </p>
-            <button
-              type="button"
-              onClick={output && !output.passed ? resetLab : runExperiment}
-              disabled={running || output?.passed}
-              className="w-full rounded-xl border border-amber-300 bg-amber-300 px-5 py-2.5 font-mono text-xs font-bold text-stone-950 transition hover:bg-amber-200 focus:outline-none focus-visible:ring-2 focus-visible:ring-amber-300 focus-visible:ring-offset-2 focus-visible:ring-offset-stone-900 disabled:cursor-wait disabled:border-stone-600 disabled:bg-stone-700 disabled:text-stone-400 sm:w-auto sm:shrink-0"
-            >
-              {operation === 'booting'
-                ? 'Starting PostgreSQL…'
-                : operation === 'preparing'
-                ? 'Preparing lesson…'
-                : operation === 'lesson'
-                ? 'Running…'
-                : output?.passed
-                ? 'Done'
-                : output
-                ? 'Restore lesson to retry'
-                : experiment.actionLabel}
-            </button>
-          </div>
-          <AccessPath path={experiment.path} label={lesson.pathLabel} />
-          <div aria-live="polite" aria-busy={running}>
-            <GuidedSqlNotebook
-              experiment={experiment}
-              output={output}
-              operation={operation}
-            />
-          </div>
-        </div>
-
-        {output?.passed && (
-          <div className="rounded-2xl border border-emerald-200 bg-emerald-50/70 p-5 dark:border-emerald-900 dark:bg-emerald-950/25">
-            <p className="m-0 font-mono text-[10px] font-semibold uppercase tracking-[0.16em] text-emerald-700 dark:text-emerald-300">
-              What this shows
-            </p>
-            <p className="mb-0 mt-2 text-sm leading-6 text-stone-800 dark:text-stone-200">
-              {experiment.explanation}
-            </p>
-            <p className="mb-0 mt-3 border-l-2 border-emerald-500 pl-3 text-sm font-semibold leading-6 text-stone-900 dark:text-white">
-              {experiment.takeaway}
-            </p>
-          </div>
-        )}
-
-        <div className="flex flex-col-reverse gap-3 border-t border-stone-200 pt-5 dark:border-stone-800 sm:flex-row sm:items-center sm:justify-between">
-          <button
-            type="button"
-            onClick={showPrevious}
-            disabled={running || activeIndex === 0}
-            className="rounded-lg border border-stone-300 px-4 py-2 text-sm font-semibold text-stone-600 hover:border-stone-500 hover:text-stone-950 disabled:cursor-not-allowed disabled:opacity-40 dark:border-stone-700 dark:text-stone-300 dark:hover:text-white"
-          >
-            ← Back
-          </button>
-          {activeIndex < experiments.length - 1 ? (
-            <button
-              type="button"
-              onClick={showNext}
-              disabled={running || !output?.passed}
-              className="rounded-lg border border-stone-900 bg-stone-900 px-4 py-2 text-sm font-semibold text-white hover:bg-stone-700 disabled:cursor-not-allowed disabled:border-stone-200 disabled:bg-stone-100 disabled:text-stone-400 dark:border-white dark:bg-white dark:text-stone-900 dark:hover:bg-stone-200 dark:disabled:border-stone-700 dark:disabled:bg-stone-800 dark:disabled:text-stone-500"
-            >
-              Continue: {experiments[activeIndex + 1].title} →
-            </button>
-          ) : output?.passed ? (
-            nextLesson ? (
-              <Link
-                href={nextLesson.href}
-                className="rounded-lg border border-stone-900 bg-stone-900 px-4 py-2 text-sm font-semibold text-white no-underline hover:bg-stone-700 dark:border-white dark:bg-white dark:text-stone-900 dark:hover:bg-stone-200"
-              >
-                Next lesson: {nextLesson.title} →
-              </Link>
-            ) : (
-              <p className="m-0 text-sm font-semibold text-emerald-700 dark:text-emerald-300">
-                Lesson complete.
-              </p>
-            )
-          ) : (
-            <p className="m-0 text-sm text-stone-500 dark:text-stone-400">
-              Run this step to continue.
-            </p>
-          )}
-        </div>
-      </section>
-
-      <div className="border-t border-stone-200 bg-stone-50/70 dark:border-stone-800 dark:bg-stone-950/30">
-        <button
-          type="button"
-          aria-expanded={playgroundOpen}
-          aria-controls="postgres-sql-playground"
-          onClick={() => setPlaygroundOpen((open) => !open)}
-          disabled={running}
-          className="flex w-full items-center justify-between gap-4 px-5 py-5 text-left hover:bg-amber-50/60 disabled:cursor-wait disabled:opacity-60 dark:hover:bg-stone-800/50 sm:px-7"
-        >
-          <span>
-            <span className="block font-display text-lg text-stone-950 dark:text-white">
-              Try your own SQL
-            </span>
-            <span className="mt-1 block text-sm leading-6 text-stone-600 dark:text-stone-400">
-              Available at every step · choose a role and run your own SQL
-            </span>
-          </span>
-          <span aria-hidden="true" className="font-mono text-lg text-stone-500">
-            {playgroundOpen ? '−' : '+'}
-          </span>
-        </button>
-
-        {playgroundOpen && (
-          <section
-            id="postgres-sql-playground"
-            className="grid gap-4 border-t border-stone-200 p-5 dark:border-stone-800 sm:p-7"
-            aria-label="SQL playground"
-          >
-            <p className="m-0 text-sm leading-6 text-stone-600 dark:text-stone-300">
-              Each run uses a temporary copy of the latest executed lesson
-              state, so your changes are discarded and the guided steps stay
-              intact. The selected role is the starting identity; your SQL may
-              change it inside that disposable run.
-            </p>
-            <div className="flex flex-col gap-2 sm:flex-row sm:items-end sm:justify-between">
-              <label className="grid gap-1.5 text-sm font-semibold text-stone-700 dark:text-stone-200">
-                Start as
-                <select
-                  value={playgroundRole}
-                  onChange={(event) => setPlaygroundRole(event.target.value)}
-                  disabled={running}
-                  className="min-w-44 rounded-lg border border-stone-300 bg-white px-3 py-2 font-mono text-sm text-stone-900 dark:border-stone-700 dark:bg-stone-900 dark:text-white"
-                >
-                  {lesson.playgroundRoles.map((role) => (
-                    <option key={role.value} value={role.value}>
-                      {role.label}
-                    </option>
-                  ))}
-                </select>
-              </label>
-              <div className="flex flex-wrap gap-2">
-                <button
-                  type="button"
-                  onClick={() => setPlaygroundSql(starterSql)}
-                  disabled={running}
-                  className="rounded-lg border border-stone-300 px-3 py-2 text-xs font-semibold text-stone-600 hover:border-amber-400 hover:text-stone-950 dark:border-stone-700 dark:text-stone-300 dark:hover:text-white"
-                >
-                  Who am I?
-                </button>
-                <button
-                  type="button"
-                  onClick={() =>
-                    setPlaygroundSql('SELECT * FROM app.orders ORDER BY id;')
-                  }
-                  disabled={running}
-                  className="rounded-lg border border-stone-300 px-3 py-2 text-xs font-semibold text-stone-600 hover:border-amber-400 hover:text-stone-950 dark:border-stone-700 dark:text-stone-300 dark:hover:text-white"
-                >
-                  Read orders
-                </button>
-              </div>
-            </div>
-            <SqlEditor
-              value={playgroundSql}
-              onChange={setPlaygroundSql}
-              disabled={running}
-            />
-            <div className="flex justify-end">
-              <button
-                type="button"
-                onClick={runPlaygroundSql}
-                disabled={running || playgroundSql.trim() === ''}
-                className="w-full rounded-xl border border-amber-300 bg-amber-300 px-5 py-2.5 font-mono text-xs font-bold text-stone-950 hover:bg-amber-200 disabled:cursor-wait disabled:border-stone-300 disabled:bg-stone-200 disabled:text-stone-500 sm:w-auto"
-              >
-                {operation === 'playground'
-                  ? 'Running SQL…'
-                  : running
-                  ? 'PostgreSQL is busy…'
-                  : 'Run SQL'}
-              </button>
-            </div>
-            <div
-              className="rounded-2xl border border-stone-700 bg-stone-950 p-5 text-stone-200"
-              aria-live="polite"
-              aria-busy={operation === 'playground'}
-            >
-              <p className="mb-4 mt-0 font-mono text-[10px] font-semibold uppercase tracking-[0.16em] text-stone-500">
-                Playground output
-              </p>
-              <SqlOutput
-                output={playgroundOutput}
-                operation={operation}
-                kind="playground"
-              />
-            </div>
-          </section>
-        )}
-      </div>
-
-      <details className="border-t border-stone-200 px-5 py-4 dark:border-stone-800 sm:px-7">
-        <summary className="cursor-pointer text-sm font-semibold text-stone-600 hover:text-stone-950 dark:text-stone-300 dark:hover:text-white">
-          How “Run as” works in the browser
-        </summary>
-        <p className="mb-0 mt-3 text-sm leading-6 text-stone-500 dark:text-stone-400">
-          The selector sets the starting session authorization inside this
-          isolated browser database, so ordinary permission checks and{' '}
-          <code>session_user</code>/<code>current_user</code> initially reflect
-          the chosen role. The scratchpad accepts arbitrary SQL, so it is not an
-          authentication boundary and SQL may change that identity. It does not
-          test passwords, <code>CONNECT</code>, <code>LOGIN</code>, or{' '}
-          <code>pg_hba.conf</code>.
-        </p>
-      </details>
-
-      <details className="border-t border-stone-200 px-5 py-4 dark:border-stone-800 sm:px-7">
-        <summary className="cursor-pointer text-sm font-semibold text-stone-600 hover:text-stone-950 dark:text-stone-300 dark:hover:text-white">
-          Show the lesson setup SQL
-        </summary>
-        <p className="mb-4 mt-3 text-sm leading-6 text-stone-500 dark:text-stone-400">
-          The lesson prepares {lesson.setupSummary}. It runs once before the
-          first query.
-        </p>
-        <div className="overflow-hidden rounded-2xl border border-stone-700">
-          <SqlSource
-            statements={lesson.setupStatements}
-            label="Lesson setup SQL"
-            maxHeight="max-h-[24rem]"
-          />
-        </div>
-      </details>
-    </div>
-  )
-}
-
-export function PostgresPermissionLab() {
-  return <AccessStoryLab />
-}
-
-export function PostgresRoleHierarchyLab() {
-  return <PostgresLesson lesson={hierarchyLesson} />
-}
